@@ -1,41 +1,71 @@
 module.exports = {
-  'env': {
-    'node': true,
-    'mocha': true,
-    'es6': true
+  env: {
+    node: true,
+    mocha: true,
+    es6: true,
   },
-  'extends': 'eslint:recommended',
-  'parserOptions': {
-    'sourceType': 'module',
-      "ecmaVersion": 8
+  ignorePatterns: ['src/test/*.*', 'examples/**/*'],
+  overrides: [],
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+    'prettier', // This should be the last entry.
+  ],
+  parser: '@typescript-eslint/parser',
+  plugins: ['@typescript-eslint', 'simple-import-sort'],
+  parserOptions: {
+    sourceType: 'module',
+    ecmaVersion: 8,
   },
-  'rules': {
-    'indent': [
+  rules: {
+    'no-console': ['error'],
+    // "no-var": ["error"],
+    'comma-dangle': 0,
+    curly: ['error'],
+    'prefer-const': 0,
+    'no-template-curly-in-string': 'error',
+    // "quotes": ["error", "double"],
+    'comma-spacing': 0, // ["error", { before: false, after: true }],
+    'semi-spacing': 0, // ["warn", { before: false, after: true }],
+    'space-before-blocks': 0, // ["warn", "always"],
+    'switch-colon-spacing': ['warn', { after: true, before: false }],
+    'keyword-spacing': 0, // ["warn", { before: true, after: true }],
+    'template-curly-spacing': 0, // ["error", "never"],
+    'rest-spread-spacing': 0, // ["error", "never"],
+    'no-multi-spaces': 0, // ["warn", { ignoreEOLComments: false }],
+
+    'simple-import-sort/imports': 'error',
+    'simple-import-sort/exports': 'error',
+    indent: 'off',
+    'linebreak-style': ['error', 'unix'],
+    semi: ['error', 'never'],
+    'spaced-comment': [
       'error',
-      2,
+      'always',
       {
-        'FunctionDeclaration': { 'parameters': 'first' },
-        'FunctionExpression': { 'parameters': 'first' },
-        'CallExpression': { 'arguments': 'first' },
-        'ArrayExpression': 'first',
-        'ObjectExpression': 'first'
-      }
+        line: {
+          markers: ['/'],
+          exceptions: ['-', '+'],
+        },
+        block: {
+          markers: ['!'],
+          exceptions: ['*'],
+          balanced: true,
+        },
+      },
     ],
-    'linebreak-style': [
-      'error',
-      (process.platform === 'win32' ? 'windows' : 'unix') // all windows platforms are denoted by win32
-    ],
-    'semi': ['error', 'never'],
-    'spaced-comment': ['error', 'always', {
-      'line': {
-        'markers': ['/'],
-        'exceptions': ['-', '+']
-    },
-    'block': {
-      'markers': ['!'],
-      'exceptions': ['*'],
-      'balanced': true
-    }
-    }]
-  }
+    '@typescript-eslint/no-explicit-any': ['warn'],
+
+    '@typescript-eslint/prefer-optional-chain': 0, // ["warn"],
+    'no-empty-function': 0,
+    '@typescript-eslint/no-empty-function': 0, // ["warn"],
+    '@typescript-eslint/no-var-requires': 0,
+    '@typescript-eslint/no-this-alias': 0,
+    '@typescript-eslint/no-empty-interface': ['warn'],
+
+    '@typescript-eslint/no-array-constructor': ['off'],
+
+    'no-extra-parens': 0,
+    '@typescript-eslint/no-extra-parens': 0,
+  },
 }
