@@ -17,28 +17,25 @@
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-bucketname and my-objectname
 // are dummy values, please replace them with original values.
 const os = require('os')
-const splitFile = require("split-file")
+const splitFile = require('split-file')
 const fs = require('fs')
 
-var Minio = require("../dist/main/minio")
-var Helpers = require("../dist/main/helpers")
-
+var Minio = require('../dist/main/minio')
+var Helpers = require('../dist/main/helpers')
 
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
 const oneMB = 1024 * 1024
-
 
 // Create a bucket prior to running: mc mb local/source-bucket
 function sampleRunComposeObject() {
   var tmpDir = os.tmpdir()
 
-
-  const bucketName = "source-bucket"
+  const bucketName = 'source-bucket'
   // generate 100 MB buffer and write to a file.
   var local100mbFileToBeSplitAndComposed = Buffer.alloc(100 * oneMB, 0)
 
@@ -47,94 +44,94 @@ function sampleRunComposeObject() {
   var fileToSplit = `${tmpSubDir}/${composedObjName}`
   let partObjNameList = []
 
-  fs.mkdir(tmpSubDir, { recursive: true }, function(err) {
+  fs.mkdir(tmpSubDir, { recursive: true }, function (err) {
     if (err) {
       console.log(err)
     } else {
-      console.log("New Temp directory successfully created.")
+      console.log('New Temp directory successfully created.')
     }
   })
 
-
   try {
     fs.writeFileSync(fileToSplit, local100mbFileToBeSplitAndComposed)
-    console.log("Written 100 MB File ")
+    console.log('Written 100 MB File ')
     // 100 MB split into 26 MB part size. ( just to test unequal parts ). But change as required.
 
-    splitFile.splitFileBySize(fileToSplit, (26 * oneMB))
+    splitFile
+      .splitFileBySize(fileToSplit, 26 * oneMB)
       .then((names) => {
-        console.log("Split and write 100 MB File(s) ", names)
+        console.log('Split and write 100 MB File(s) ', names)
         const putPartRequests = names.map((partFileName) => {
-          const partObjName = partFileName.slice((tmpSubDir + "/").length)
+          const partObjName = partFileName.slice((tmpSubDir + '/').length)
           partObjNameList.push(partObjName)
           return s3Client.fPutObject(bucketName, partObjName, partFileName, {})
         })
 
-        Promise.all(putPartRequests).then(() => {
-          console.log("Uploaded part Files: ", names)
-          const sourcePartObjList = partObjNameList.map((partObjName) => {
-            return new Helpers.CopySourceOptions({
-              Bucket: bucketName,
-              Object: partObjName
-            })
-          })
-
-          const destObjConfig = new Helpers.CopyDestinationOptions({
-            Bucket: bucketName,
-            Object: composedObjName
-          })
-
-          s3Client.composeObject(destObjConfig, sourcePartObjList).then(() => {
-            console.log("Composed to a single file: ", composedObjName)
-
-            /** Begin Clean up ***/
-            // To verify that the parts are uploaded properly, comment the below code blocks and verify
+        Promise.all(putPartRequests)
+          .then(() => {
+            console.log('Uploaded part Files: ', names)
             const sourcePartObjList = partObjNameList.map((partObjName) => {
-              return s3Client.removeObject(bucketName, partObjName)
+              return new Helpers.CopySourceOptions({
+                Bucket: bucketName,
+                Object: partObjName,
+              })
             })
 
-            Promise.all(sourcePartObjList).then(() => {
-              console.log("Removed source parts: ")
+            const destObjConfig = new Helpers.CopyDestinationOptions({
+              Bucket: bucketName,
+              Object: composedObjName,
+            })
 
-              // Uncomment to remove the composed object itself. commented for verification.
-              /*
+            s3Client
+              .composeObject(destObjConfig, sourcePartObjList)
+              .then(() => {
+                console.log('Composed to a single file: ', composedObjName)
+
+                /** Begin Clean up ***/
+                // To verify that the parts are uploaded properly, comment the below code blocks and verify
+                const sourcePartObjList = partObjNameList.map((partObjName) => {
+                  return s3Client.removeObject(bucketName, partObjName)
+                })
+
+                Promise.all(sourcePartObjList)
+                  .then(() => {
+                    console.log('Removed source parts: ')
+
+                    // Uncomment to remove the composed object itself. commented for verification.
+                    /*
               s3Client.removeObject(bucketName, composedObjName).then(()=>{
                 console.log("Clean up: Removed the composed Object ")
               }).catch(()=>{
                 console.log("Error removing composed object", er)
               })
               */
+                  })
+                  .catch((er) => {
+                    console.log('Error removing parts used in composing', er)
+                  })
 
-            }).catch(er => {
-              console.log("Error removing parts used in composing", er)
-            })
+                /** End Clean up **/
 
-            /** End Clean up **/
-
-            // Clean up generated parts locally
-            Helpers.removeDirAndFiles(tmpSubDir)
-            console.log("Clean up temp parts directory : ")
-
-          }).catch((e) => {
-            console.log("Error Composing parts into an object", e)
+                // Clean up generated parts locally
+                Helpers.removeDirAndFiles(tmpSubDir)
+                console.log('Clean up temp parts directory : ')
+              })
+              .catch((e) => {
+                console.log('Error Composing parts into an object', e)
+              })
           })
-
-        }).catch(e => {
-          console.log("Error Uploading parts ", e)
-        })
-
-
+          .catch((e) => {
+            console.log('Error Uploading parts ', e)
+          })
       })
       .catch((e) => {
         // this is a client error not related to compose object
-        console.log("Error Splitting files into parts ", e)
+        console.log('Error Splitting files into parts ', e)
       })
   } catch (err) {
     // this is a client error not related to compose object
-    console.log("Error Creating local files ", err)
+    console.log('Error Creating local files ', err)
   }
-
-
 }
 
 sampleRunComposeObject()
