@@ -14,23 +14,21 @@
  * limitations under the License.
  */
 
- // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY and my-bucketname are
- // dummy values, please replace them with original values.
-
+// Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY and my-bucketname are
+// dummy values, please replace them with original values.
 
 var Minio = require('minio')
 
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
-s3Client.getBucketNotification('my-bucketname', function(e, bucketNotification) {
+s3Client.getBucketNotification('my-bucketname', function (e, bucketNotification) {
   if (e) {
     return console.log(e)
   }
   console.log(bucketNotification)
-  console.log("True")
+  console.log('True')
 })
-
