@@ -22,12 +22,12 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
 // Retrieves the bucket policy and logs it to the console.
 s3Client.getBucketPolicy('testbucket', (err, policy) => {
-	if (err) throw err
+  if (err) throw err
 
-	console.log(`Bucket policy: ${policy}`)
+  console.log(`Bucket policy: ${policy}`)
 })
