@@ -1,4 +1,3 @@
-
 /*
  * MinIO Javascript Library for Amazon S3 Compatible Cloud Storage, (C) 2021 MinIO, Inc.
  *
@@ -23,13 +22,12 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
-
-s3Client.getBucketTagging('bucketname', function (err, tagsList){
+s3Client.getBucketTagging('bucketname', function (err, tagsList) {
   if (err) {
     return console.log(err)
   }
-  console.log("Success", tagsList)
+  console.log('Success', tagsList)
 })
