@@ -22,14 +22,13 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
-
-s3Client.getBucketVersioning('my-bucketname', function (err,res){
+s3Client.getBucketVersioning('my-bucketname', function (err, res) {
   if (err) {
     return console.log(err)
   }
   console.log(res)
-  console.log("Success")
+  console.log('Success')
 })
