@@ -22,11 +22,10 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
-
-s3Client.getObjectLockConfig('my-bucketname', function (err,lockConfig){
+s3Client.getObjectLockConfig('my-bucketname', function (err, lockConfig) {
   if (err) {
     return console.log(err)
   }
