@@ -1,4 +1,3 @@
-
 /*
  * MinIO Javascript Library for Amazon S3 Compatible Cloud Storage, (C) 2021 MinIO, Inc.
  *
@@ -23,22 +22,21 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
-
-s3Client.getObjectTagging('bucketname', "objectName", function (err, tagsList){
+s3Client.getObjectTagging('bucketname', 'objectName', function (err, tagsList) {
   if (err) {
     return console.log(err)
   }
-  console.log("Success",tagsList)
+  console.log('Success', tagsList)
 })
 
 //Get tags on a version of an object.
 
-s3Client.getObjectTagging('bucketname', "objectName", { versionId: "" }, function (err, tagsList){
+s3Client.getObjectTagging('bucketname', 'objectName', { versionId: '' }, function (err, tagsList) {
   if (err) {
     return console.log(err)
   }
-  console.log("Success",tagsList)
+  console.log('Success', tagsList)
 })
