@@ -14,21 +14,20 @@
  * limitations under the License.
  */
 
-
-  // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-bucketname and
-  // my-objectname are dummy values, please replace them with original values.
+// Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-bucketname and
+// my-objectname are dummy values, please replace them with original values.
 
 var Minio = require('minio')
 
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 // Remove a partially uploaded object name my-objectname.
-s3Client.removeIncompleteUpload('my-bucketname', 'my-objectname', function(e) {
+s3Client.removeIncompleteUpload('my-bucketname', 'my-objectname', function (e) {
   if (e) {
     return console.log(e)
   }
-  console.log("Success")
+  console.log('Success')
 })
