@@ -17,7 +17,6 @@
 // Note: YOUR-ACCESSKEYID and YOUR-SECRETACCESSKEY are dummy values, please
 // replace them with original values.
 
-
 var Minio = require('minio')
 
 var s3Client = new Minio.Client({
@@ -25,7 +24,7 @@ var s3Client = new Minio.Client({
   port: 9000,
   useSSL: false,
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
 const objectsList = []
@@ -33,11 +32,9 @@ const bucketName = 'my-bucket'
 const prefix = 'my-prefix'
 const recursive = false
 
-
-
-function removeObjects (bucketName, prefix, recursive, includeVersion) {
-// List all object paths in bucket
-  var objectsStream = s3Client.listObjects(bucketName, prefix, recursive, {IncludeVersion: includeVersion})
+function removeObjects(bucketName, prefix, recursive, includeVersion) {
+  // List all object paths in bucket
+  var objectsStream = s3Client.listObjects(bucketName, prefix, recursive, { IncludeVersion: includeVersion })
 
   objectsStream.on('data', function (obj) {
     if (includeVersion) {
@@ -56,7 +53,7 @@ function removeObjects (bucketName, prefix, recursive, includeVersion) {
       if (e) {
         return console.log(e)
       }
-      console.log("Success")
+      console.log('Success')
     })
   })
 }
@@ -64,21 +61,19 @@ function removeObjects (bucketName, prefix, recursive, includeVersion) {
 removeObjects(bucketName, prefix, recursive, true) // Versioned objects of a bucket to be deleted.
 removeObjects(bucketName, prefix, recursive, false) // Normal objects of a bucket to be deleted.
 
-
 // Delete Multiple objects and respective versions.
 function removeObjectsMultipleVersions() {
-
   const deleteList = [
-    {versionId: '03ed08e1-34ff-4465-91ed-ba50c1e80f39', name: 'prefix-1/out.json.gz'},
-    {versionId: "35517ae1-18cb-4a21-9551-867f53a10cfe", name:"dir1/dir2/test.pdf"},
-    {versionId: "3053f564-9aea-4a59-88f0-7f25d6320a2c", name:"dir1/dir2/test.pdf"}
+    { versionId: '03ed08e1-34ff-4465-91ed-ba50c1e80f39', name: 'prefix-1/out.json.gz' },
+    { versionId: '35517ae1-18cb-4a21-9551-867f53a10cfe', name: 'dir1/dir2/test.pdf' },
+    { versionId: '3053f564-9aea-4a59-88f0-7f25d6320a2c', name: 'dir1/dir2/test.pdf' },
   ]
 
-  s3Client.removeObjects("my-bucket", deleteList,  function (e) {
+  s3Client.removeObjects('my-bucket', deleteList, function (e) {
     if (e) {
       return console.log(e)
     }
-    console.log("Successfully deleted..")
+    console.log('Successfully deleted..')
   })
 }
 removeObjectsMultipleVersions()
