@@ -22,13 +22,13 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
 //Example to reset/remove object lock config.
-s3Client.setObjectLockConfig('my-bucketname', {}, function (err){
+s3Client.setObjectLockConfig('my-bucketname', {}, function (err) {
   if (err) {
     return console.log(err)
   }
-  console.log("Success")
+  console.log('Success')
 })
