@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-bucketname
 // and my-objectname are dummy values, please replace them with original values.
 
@@ -23,19 +22,18 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
 // Enable S3 transfer accelerate endpoint.
-s3Client.SetS3TransferAccelerate("s3-accelerate.amazonaws.com")
+s3Client.SetS3TransferAccelerate('s3-accelerate.amazonaws.com')
 
 // Upload a buffer
 var buf = new Buffer(10)
 buf.fill('a')
-s3Client.putObject('my-bucketname', 'my-objectname2', buf, 'application/octet-stream', function(e) {
+s3Client.putObject('my-bucketname', 'my-objectname2', buf, 'application/octet-stream', function (e) {
   if (e) {
     return console.log(e.message)
   }
-  console.log("Success")
+  console.log('Success')
 })
-
