@@ -1,4 +1,3 @@
-
 /*
  * MinIO Javascript Library for Amazon S3 Compatible Cloud Storage, (C) 2021 MinIO, Inc.
  *
@@ -23,17 +22,16 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
 //Apply default encryption.
-s3Client.setBucketEncryption("my-bucket", function (error){
+s3Client.setBucketEncryption('my-bucket', function (error) {
   if (error) {
     return console.log(error)
   }
-  console.log("Success")
+  console.log('Success')
 })
-
 
 //Set Encryption Rule. Only one rule is allowed.
 
@@ -43,23 +41,21 @@ s3Client.setBucketEncryption("my-bucket", function (error){
  */
 
 var encryptionConfig = {
-  Rule:[
+  Rule: [
     {
       ApplyServerSideEncryptionByDefault: {
-        SSEAlgorithm:"AES256"
-      }
-    }
-  ]
-
+        SSEAlgorithm: 'AES256',
+      },
+    },
+  ],
 }
 
-s3Client.setBucketEncryption("my-bucket", encryptionConfig, function (error){
+s3Client.setBucketEncryption('my-bucket', encryptionConfig, function (error) {
   if (error) {
     return console.log(error)
   }
-  console.log("Success")
+  console.log('Success')
 })
-
 
 /**
  * KMS ID based SSE Encryption
@@ -90,19 +86,19 @@ s3Client.setBucketEncryption("my-bucket", encryptionConfig, function (error){
  */
 
 const kmsIdEncryptionConfig = {
-  Rule:[
+  Rule: [
     {
       ApplyServerSideEncryptionByDefault: {
-        KMSMasterKeyID:'my-minio-key', //as per env value
-        SSEAlgorithm:"aws:kms" // this is important
-      }
-    }
-  ]
+        KMSMasterKeyID: 'my-minio-key', //as per env value
+        SSEAlgorithm: 'aws:kms', // this is important
+      },
+    },
+  ],
 }
 
-s3Client.setBucketEncryption("my-bucket", kmsIdEncryptionConfig, function (error){
+s3Client.setBucketEncryption('my-bucket', kmsIdEncryptionConfig, function (error) {
   if (error) {
     return console.log(error)
   }
-  console.log("Success")
+  console.log('Success')
 })
