@@ -22,54 +22,54 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
-
-const lifecycleConfig= {
-  Rule: [{
-    "ID": "Expiration Days Rule",
-    "Status": "Enabled",
-    "Filter": {
-      "Prefix":"",
+const lifecycleConfig = {
+  Rule: [
+    {
+      ID: 'Expiration Days Rule',
+      Status: 'Enabled',
+      Filter: {
+        Prefix: '',
+      },
+      Expiration: {
+        Days: '3650',
+      },
     },
-    "Expiration": {
-      "Days": "3650"
-    }
-  }
-  ]
+  ],
 }
 
-s3Client.setBucketLifecycle('bucketname',lifecycleConfig, function (err) {
+s3Client.setBucketLifecycle('bucketname', lifecycleConfig, function (err) {
   if (err) {
     return console.log(err)
   }
-  console.log("Success")
+  console.log('Success')
 })
 
 //Example to demonstrate Expiration Date
 const expirationDate = new Date()
 expirationDate.setDate(expirationDate.getDate() + 1) // Expire next day.
-expirationDate.setUTCHours(0,0,0,0)//Should be start of the day.(midnight)
+expirationDate.setUTCHours(0, 0, 0, 0) //Should be start of the day.(midnight)
 
-const lifecycleConfigWithExpirationDate= {
-  Rule: [{
-    "ID": "Expiration Date Rule",
-    "Status": "Enabled",
-    "Filter": {
-      "Prefix":"",
+const lifecycleConfigWithExpirationDate = {
+  Rule: [
+    {
+      ID: 'Expiration Date Rule',
+      Status: 'Enabled',
+      Filter: {
+        Prefix: '',
+      },
+      Expiration: {
+        Date: expirationDate.toISOString(),
+      },
     },
-    "Expiration": {
-      Date:expirationDate.toISOString()
-    }
-  }
-  ]
+  ],
 }
 
-s3Client.setBucketLifecycle('bucketname',lifecycleConfigWithExpirationDate, function (err) {
+s3Client.setBucketLifecycle('bucketname', lifecycleConfigWithExpirationDate, function (err) {
   if (err) {
     return console.log(err)
   }
-  console.log("Success")
+  console.log('Success')
 })
-
