@@ -1,4 +1,3 @@
-
 /*
  * MinIO Javascript Library for Amazon S3 Compatible Cloud Storage, (C) 2021 MinIO, Inc.
  *
@@ -23,14 +22,14 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
-var versioningStateConfig = {Status:"Enabled"}
+var versioningStateConfig = { Status: 'Enabled' }
 
-s3Client.setBucketVersioning("my-bucket", versioningStateConfig, function (error){
+s3Client.setBucketVersioning('my-bucket', versioningStateConfig, function (error) {
   if (error) {
     return console.log(error)
   }
-  console.log("Success")
+  console.log('Success')
 })
