@@ -22,11 +22,11 @@ var Minio = require('minio')
 var s3Client = new Minio.Client({
   endPoint: 's3.amazonaws.com',
   accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
 //Set legal hold config of an object.
-s3Client.setObjectLegalHold('bucketName', 'objectName', {status:"ON"}, function(err, res) {
+s3Client.setObjectLegalHold('bucketName', 'objectName', { status: 'ON' }, function (err, res) {
   if (err) {
     return console.log('Unable to set legal hold config for the object', err)
   }
@@ -34,9 +34,14 @@ s3Client.setObjectLegalHold('bucketName', 'objectName', {status:"ON"}, function(
 })
 
 //Set legal hold config of an object with versionId.
-s3Client.setObjectLegalHold('bucketName', 'objectName', { status:"ON", versionId:'my-obj-version-uuid' }, function(err, res) {
-  if (err) {
-    return console.log('Unable to set legal hold config for the object version', err)
-  }
-  console.log('Success')
-})
+s3Client.setObjectLegalHold(
+  'bucketName',
+  'objectName',
+  { status: 'ON', versionId: 'my-obj-version-uuid' },
+  function (err, res) {
+    if (err) {
+      return console.log('Unable to set legal hold config for the object version', err)
+    }
+    console.log('Success')
+  },
+)
