@@ -1,15 +1,11 @@
-import Credentials from "./Credentials"
+import Credentials from './Credentials'
 
 class CredentialProvider {
-  constructor({
-    accessKey,
-    secretKey,
-    sessionToken
-  }) {
+  constructor({ accessKey, secretKey, sessionToken }) {
     this.credentials = new Credentials({
       accessKey,
       secretKey,
-      sessionToken
+      sessionToken,
     })
   }
 
@@ -21,10 +17,9 @@ class CredentialProvider {
     if (credentials instanceof Credentials) {
       this.credentials = credentials
     } else {
-      throw new Error("Unable to set Credentials . it should be an instance of Credentials class")
+      throw new Error('Unable to set Credentials . it should be an instance of Credentials class')
     }
   }
-
 
   setAccessKey(accessKey) {
     this.credentials.setAccessKey(accessKey)
@@ -49,8 +44,6 @@ class CredentialProvider {
   getSessionToken() {
     return this.credentials.getSessionToken()
   }
-
-
 }
 
 export default CredentialProvider
