@@ -16,18 +16,12 @@
 
 import Stream from 'stream'
 
+import * as errors from './errors.js'
+import { isBoolean, isNumber, isString, isValidBucketName, isValidPrefix, pipesetup, uriEscape } from './helpers.js'
 import * as transformers from './transformers'
 
-import * as errors from './errors.js'
-
-import { isValidPrefix, isValidBucketName, uriEscape,
-  isBoolean, isNumber, isString,
-  pipesetup } from './helpers.js'
-
 export default class extensions {
-
-  constructor(client)
-  {
+  constructor(client) {
     this.client = client
   }
 
@@ -49,9 +43,15 @@ export default class extensions {
   //   * `obj.metadata` _object_: metadata of the object
 
   listObjectsV2WithMetadata(bucketName, prefix, recursive, startAfter) {
-    if (prefix === undefined) prefix = ''
-    if (recursive === undefined) recursive = false
-    if (startAfter === undefined) startAfter = ''
+    if (prefix === undefined) {
+      prefix = ''
+    }
+    if (recursive === undefined) {
+      recursive = false
+    }
+    if (startAfter === undefined) {
+      startAfter = ''
+    }
     if (!isValidBucketName(bucketName)) {
       throw new errors.InvalidBucketNameError('Invalid bucket name: ' + bucketName)
     }
@@ -72,18 +72,20 @@ export default class extensions {
     var continuationToken = ''
     var objects = []
     var ended = false
-    var readStream = Stream.Readable({objectMode: true})
+    var readStream = Stream.Readable({ objectMode: true })
     readStream._read = () => {
       // push one object per _read()
       if (objects.length) {
         readStream.push(objects.shift())
         return
       }
-      if (ended) return readStream.push(null)
+      if (ended) {
+        return readStream.push(null)
+      }
       // if there are no objects to push do query for the next batch of objects
       this.listObjectsV2WithMetadataQuery(bucketName, prefix, continuationToken, delimiter, 1000, startAfter)
-        .on('error', e => readStream.emit('error', e))
-        .on('data', result => {
+        .on('error', (e) => readStream.emit('error', e))
+        .on('data', (result) => {
           if (result.isTruncated) {
             continuationToken = result.nextContinuationToken
           } else {
@@ -159,8 +161,10 @@ export default class extensions {
     }
     var method = 'GET'
     var transformer = transformers.getListObjectsV2WithMetadataTransformer()
-    this.client.makeRequest({method, bucketName, query}, '', [200], '', true, (e, response) => {
-      if (e) return transformer.emit('error', e)
+    this.client.makeRequest({ method, bucketName, query }, '', [200], '', true, (e, response) => {
+      if (e) {
+        return transformer.emit('error', e)
+      }
       pipesetup(response, transformer)
     })
     return transformer
