@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-import { Transform } from 'stream'
 import Crypto from 'crypto'
 import * as querystring from 'query-string'
-import { getVersionId, sanitizeETag } from "./helpers"
+import { Transform } from 'stream'
+
+import { getVersionId, sanitizeETag } from './helpers'
 
 // We extend Transform because Writable does not implement ._flush().
 export default class ObjectUploader extends Transform {
@@ -52,7 +53,7 @@ export default class ObjectUploader extends Transform {
     this.id = null
 
     // Handle errors.
-    this.on('error', err => {
+    this.on('error', (err) => {
       callback(err)
     })
   }
@@ -60,7 +61,7 @@ export default class ObjectUploader extends Transform {
   _transform(chunk, encoding, callback) {
     this.emptyStream = false
     let method = 'PUT'
-    let headers = {'Content-Length': chunk.length}
+    let headers = { 'Content-Length': chunk.length }
     let md5digest = ''
 
     // Calculate and set Content-MD5 header if SHA256 is not set.
@@ -80,14 +81,16 @@ export default class ObjectUploader extends Transform {
         headers: Object.assign({}, this.metaData, headers),
         query: '',
         bucketName: this.bucketName,
-        objectName: this.objectName
+        objectName: this.objectName,
       }
 
       this.client.makeRequest(options, chunk, [200], '', true, (err, response) => {
-        if (err) return callback(err)
+        if (err) {
+          return callback(err)
+        }
         let result = {
           etag: sanitizeETag(response.headers.etag),
-          versionId :getVersionId(response.headers)
+          versionId: getVersionId(response.headers),
         }
         // Ignore the 'data' event so that the stream closes. (nodejs stream requirement)
         response.on('data', () => {})
@@ -115,12 +118,16 @@ export default class ObjectUploader extends Transform {
 
       // Check for an incomplete previous upload.
       this.client.findUploadId(this.bucketName, this.objectName, (err, id) => {
-        if (err) return this.emit('error', err)
+        if (err) {
+          return this.emit('error', err)
+        }
 
         // If no upload ID exists, initiate a new one.
         if (!id) {
           this.client.initiateNewMultipartUpload(this.bucketName, this.objectName, this.metaData, (err, id) => {
-            if (err) return callback(err)
+            if (err) {
+              return callback(err)
+            }
 
             this.id = id
 
@@ -135,13 +142,17 @@ export default class ObjectUploader extends Transform {
 
         // Retrieve the pre-uploaded parts, if we need to resume the upload.
         this.client.listParts(this.bucketName, this.objectName, id, (err, etags) => {
-          if (err) return this.emit('error', err)
+          if (err) {
+            return this.emit('error', err)
+          }
 
           // It is possible for no parts to be already uploaded.
-          if (!etags) etags = []
+          if (!etags) {
+            etags = []
+          }
 
           // oldParts will become an object, allowing oldParts[partNumber].etag
-          this.oldParts = etags.reduce(function(prev, item) {
+          this.oldParts = etags.reduce(function (prev, item) {
             if (!prev[item.part]) {
               prev[item.part] = item
             }
@@ -164,13 +175,13 @@ export default class ObjectUploader extends Transform {
       let oldPart = this.oldParts[partNumber]
 
       // Calulcate the md5 hash, if it has not already been calculated.
-      if(!md5digest) {
+      if (!md5digest) {
         md5digest = Crypto.createHash('md5').update(chunk).digest()
       }
 
       if (oldPart && md5digest.toString('hex') === oldPart.etag) {
         // The md5 matches, the chunk has already been uploaded.
-        this.etags.push({part: partNumber, etag: oldPart.etag})
+        this.etags.push({ part: partNumber, etag: oldPart.etag })
 
         callback()
         return
@@ -180,24 +191,29 @@ export default class ObjectUploader extends Transform {
     // Write the chunk with an uploader.
     let query = querystring.stringify({
       partNumber: partNumber,
-      uploadId: this.id
+      uploadId: this.id,
     })
 
     let options = {
-      method, query, headers,
+      method,
+      query,
+      headers,
       bucketName: this.bucketName,
-      objectName: this.objectName
+      objectName: this.objectName,
     }
 
     this.client.makeRequest(options, chunk, [200], '', true, (err, response) => {
-      if (err) return callback(err)
+      if (err) {
+        return callback(err)
+      }
 
       // In order to aggregate the parts together, we need to collect the etags.
       let etag = response.headers.etag
-      if (etag)
+      if (etag) {
         etag = etag.replace(/^"/, '').replace(/"$/, '')
+      }
 
-      this.etags.push({part: partNumber, etag})
+      this.etags.push({ part: partNumber, etag })
 
       // Ignore the 'data' event so that the stream closes. (nodejs stream requirement)
       response.on('data', () => {})
@@ -210,20 +226,23 @@ export default class ObjectUploader extends Transform {
   _flush(callback) {
     if (this.emptyStream) {
       let method = 'PUT'
-      let headers = Object.assign({}, this.metaData, {'Content-Length': 0})
+      let headers = Object.assign({}, this.metaData, { 'Content-Length': 0 })
       let options = {
-        method, headers,
+        method,
+        headers,
         query: '',
         bucketName: this.bucketName,
-        objectName: this.objectName
+        objectName: this.objectName,
       }
 
       this.client.makeRequest(options, '', [200], '', true, (err, response) => {
-        if (err) return callback(err)
+        if (err) {
+          return callback(err)
+        }
 
         let result = {
           etag: sanitizeETag(response.headers.etag),
-          versionId: getVersionId(response.headers)
+          versionId: getVersionId(response.headers),
         }
 
         // Ignore the 'data' event so that the stream closes. (nodejs stream requirement)
@@ -247,17 +266,18 @@ export default class ObjectUploader extends Transform {
 
     // This is called when all of the chunks uploaded successfully, thus
     // completing the multipart upload.
-    this.client.completeMultipartUpload(this.bucketName, this.objectName, this.id,
-                                        this.etags, (err, etag) => {
-                                          if (err) return callback(err)
+    this.client.completeMultipartUpload(this.bucketName, this.objectName, this.id, this.etags, (err, etag) => {
+      if (err) {
+        return callback(err)
+      }
 
-                                          // Call our callback on the next tick to allow the streams infrastructure
-                                          // to finish what its doing before we continue.
-                                          process.nextTick(() => {
-                                            this.callback(null, etag)
-                                          })
+      // Call our callback on the next tick to allow the streams infrastructure
+      // to finish what its doing before we continue.
+      process.nextTick(() => {
+        this.callback(null, etag)
+      })
 
-                                          callback()
-                                        })
+      callback()
+    })
   }
 }
