@@ -19,15 +19,19 @@ require('source-map-support').install()
 import { assert } from 'chai'
 import Nock from 'nock'
 import Stream from 'stream'
-import * as Minio from '../../../dist/main/minio'
-import { isValidEndpoint, 
-  isValidIP, makeDateLong,
-  makeDateShort,partsRequired,
-  CopySourceOptions,
-  CopyDestinationOptions,
-  isArray,
+
+import {
   calculateEvenSplits,
+  CopyDestinationOptions,
+  CopySourceOptions,
+  isArray,
+  isValidEndpoint,
+  isValidIP,
+  makeDateLong,
+  makeDateShort,
+  partsRequired,
 } from '../../../dist/main/helpers'
+import * as Minio from '../../../dist/main/minio'
 
 var Package = require('../../../package.json')
 
@@ -44,7 +48,7 @@ describe('Helpers', () => {
   it('should fail for invalid endpoint characters', () => {
     assert.equal(isValidEndpoint('111.#2.11'), false)
   })
- 
+
   it('should make date short', () => {
     let date = new Date('2012-12-03T17:25:36.331Z')
 
@@ -56,110 +60,132 @@ describe('Helpers', () => {
     assert.equal(makeDateLong(date), '20170811T172634Z')
   })
 
-
   // Adopted from minio-go sdk
-  const oneGB =1024 * 1024 * 1024
-  const fiveGB = (5*oneGB)
+  const oneGB = 1024 * 1024 * 1024
+  const fiveGB = 5 * oneGB
 
   const OBJ_SIZES = {
-    gb1 : oneGB,
-    gb5 : fiveGB,
-    gb5p1 : fiveGB + 1,
-    gb10p1 : 2 * fiveGB + 1,
-    gb10p2 : 2 * fiveGB + 2
+    gb1: oneGB,
+    gb5: fiveGB,
+    gb5p1: fiveGB + 1,
+    gb10p1: 2 * fiveGB + 1,
+    gb10p2: 2 * fiveGB + 2,
   }
 
   const maxMultipartPutObjectSize = 1024 * 1024 * 1024 * 1024 * 5
 
-  it('Parts Required Test cases ', () =>{
-   
+  it('Parts Required Test cases ', () => {
     const expectedPartsRequiredTestCases = [
-      {value:0, expected:0},
-      {value:1, expected:1},
-      {value:fiveGB, expected:10},
-      {value:OBJ_SIZES.gb5p1, expected:10},
-      {value:2 * fiveGB, expected:20},
-      {value:OBJ_SIZES.gb10p1, expected:20},
-      {value:OBJ_SIZES.gb10p2, expected:20},
-      {value:OBJ_SIZES.gb10p1 + OBJ_SIZES.gb10p2, expected:40},
-      {value:maxMultipartPutObjectSize, expected:10000}
+      { value: 0, expected: 0 },
+      { value: 1, expected: 1 },
+      { value: fiveGB, expected: 10 },
+      { value: OBJ_SIZES.gb5p1, expected: 10 },
+      { value: 2 * fiveGB, expected: 20 },
+      { value: OBJ_SIZES.gb10p1, expected: 20 },
+      { value: OBJ_SIZES.gb10p2, expected: 20 },
+      { value: OBJ_SIZES.gb10p1 + OBJ_SIZES.gb10p2, expected: 40 },
+      { value: maxMultipartPutObjectSize, expected: 10000 },
     ]
 
-    expectedPartsRequiredTestCases.forEach((testCase)=>{
+    expectedPartsRequiredTestCases.forEach((testCase) => {
       const fnResult = partsRequired(testCase.value)
-      assert.equal(fnResult,testCase.expected)
+      assert.equal(fnResult, testCase.expected)
     })
   })
-  it("Even split of Sizes Test cases ", ()=>{
+  it('Even split of Sizes Test cases ', () => {
     // Adopted from minio-go sdk
-    const expectedSplitsTestCases =[
-      {size:0, sourceConfig:new CopySourceOptions({Start: -1}), expectedStart:null, expectedEnd:null}
-      ,{size:1, sourceConfig:new CopySourceOptions({Start: -1}), expectedStart:[undefined], expectedEnd:[NaN]}
-      ,{size:1,sourceConfig:new  CopySourceOptions({Start: 0}), expectedStart:[0], expectedEnd:[0]}
-      ,{size:OBJ_SIZES.gb1, sourceConfig:new CopySourceOptions({Start: -1}), expectedStart:[0, 536870912],expectedEnd:[536870911, 1073741823]}
-      ,{size:OBJ_SIZES.gb5, sourceConfig:new CopySourceOptions({Start: -1}),
-        expectedStart:[0, 536870912, 1073741824, 1610612736, 2147483648, 2684354560,
-                       3221225472, 3758096384, 4294967296, 4831838208],
-        expectedEnd:[536870911, 1073741823, 1610612735, 2147483647, 2684354559, 3221225471,
-                     3758096383, 4294967295, 4831838207, 5368709119]
+    const expectedSplitsTestCases = [
+      { size: 0, sourceConfig: new CopySourceOptions({ Start: -1 }), expectedStart: null, expectedEnd: null },
+      { size: 1, sourceConfig: new CopySourceOptions({ Start: -1 }), expectedStart: [undefined], expectedEnd: [NaN] },
+      { size: 1, sourceConfig: new CopySourceOptions({ Start: 0 }), expectedStart: [0], expectedEnd: [0] },
+      {
+        size: OBJ_SIZES.gb1,
+        sourceConfig: new CopySourceOptions({ Start: -1 }),
+        expectedStart: [0, 536870912],
+        expectedEnd: [536870911, 1073741823],
+      },
+      {
+        size: OBJ_SIZES.gb5,
+        sourceConfig: new CopySourceOptions({ Start: -1 }),
+        expectedStart: [
+          0, 536870912, 1073741824, 1610612736, 2147483648, 2684354560, 3221225472, 3758096384, 4294967296, 4831838208,
+        ],
+        expectedEnd: [
+          536870911, 1073741823, 1610612735, 2147483647, 2684354559, 3221225471, 3758096383, 4294967295, 4831838207,
+          5368709119,
+        ],
       },
 
       // 2 part splits
-      {size:OBJ_SIZES.gb5p1,  sourceConfig:new CopySourceOptions({Start: -1}),
-       expectedStart:[0, 536870913, 1073741825, 1610612737, 2147483649, 2684354561,
-                      3221225473, 3758096385, 4294967297, 4831838209],
-       expectedEnd:[536870912, 1073741824, 1610612736, 2147483648, 2684354560, 3221225472,
-                    3758096384, 4294967296, 4831838208, 5368709120],
+      {
+        size: OBJ_SIZES.gb5p1,
+        sourceConfig: new CopySourceOptions({ Start: -1 }),
+        expectedStart: [
+          0, 536870913, 1073741825, 1610612737, 2147483649, 2684354561, 3221225473, 3758096385, 4294967297, 4831838209,
+        ],
+        expectedEnd: [
+          536870912, 1073741824, 1610612736, 2147483648, 2684354560, 3221225472, 3758096384, 4294967296, 4831838208,
+          5368709120,
+        ],
       },
-      {size:OBJ_SIZES.gb5p1,  sourceConfig:new CopySourceOptions({Start: -1}),
-       expectedStart:[0, 536870913, 1073741825, 1610612737, 2147483649, 2684354561,
-                      3221225473, 3758096385, 4294967297, 4831838209],
-       expectedEnd:[536870912, 1073741824, 1610612736, 2147483648, 2684354560, 3221225472,
-                    3758096384, 4294967296, 4831838208, 5368709120],
+      {
+        size: OBJ_SIZES.gb5p1,
+        sourceConfig: new CopySourceOptions({ Start: -1 }),
+        expectedStart: [
+          0, 536870913, 1073741825, 1610612737, 2147483649, 2684354561, 3221225473, 3758096385, 4294967297, 4831838209,
+        ],
+        expectedEnd: [
+          536870912, 1073741824, 1610612736, 2147483648, 2684354560, 3221225472, 3758096384, 4294967296, 4831838208,
+          5368709120,
+        ],
       },
 
       // 3 part splits
-      {size:OBJ_SIZES.gb10p1,sourceConfig:new CopySourceOptions({Start: -1}),
-       expectedStart:[0, 536870913, 1073741825, 1610612737, 2147483649, 2684354561,
-                      3221225473, 3758096385, 4294967297, 4831838209, 5368709121,
-                      5905580033, 6442450945, 6979321857, 7516192769, 8053063681,
-                      8589934593, 9126805505, 9663676417, 10200547329],
-       expectedEnd:[536870912, 1073741824, 1610612736, 2147483648, 2684354560,
-                    3221225472, 3758096384, 4294967296, 4831838208, 5368709120,
-                    5905580032, 6442450944, 6979321856, 7516192768, 8053063680,
-                    8589934592, 9126805504, 9663676416, 10200547328, 10737418240],
+      {
+        size: OBJ_SIZES.gb10p1,
+        sourceConfig: new CopySourceOptions({ Start: -1 }),
+        expectedStart: [
+          0, 536870913, 1073741825, 1610612737, 2147483649, 2684354561, 3221225473, 3758096385, 4294967297, 4831838209,
+          5368709121, 5905580033, 6442450945, 6979321857, 7516192769, 8053063681, 8589934593, 9126805505, 9663676417,
+          10200547329,
+        ],
+        expectedEnd: [
+          536870912, 1073741824, 1610612736, 2147483648, 2684354560, 3221225472, 3758096384, 4294967296, 4831838208,
+          5368709120, 5905580032, 6442450944, 6979321856, 7516192768, 8053063680, 8589934592, 9126805504, 9663676416,
+          10200547328, 10737418240,
+        ],
       },
-      {size:OBJ_SIZES.gb10p2,sourceConfig:new CopySourceOptions({Start: -1}),
-       expectedStart:[0, 536870913, 1073741826, 1610612738, 2147483650, 2684354562,
-                      3221225474, 3758096386, 4294967298, 4831838210, 5368709122,
-                      5905580034, 6442450946, 6979321858, 7516192770, 8053063682,
-                      8589934594, 9126805506, 9663676418, 10200547330],
-       expectedEnd:[536870912, 1073741825, 1610612737, 2147483649, 2684354561,
-                    3221225473, 3758096385, 4294967297, 4831838209, 5368709121,
-                    5905580033, 6442450945, 6979321857, 7516192769, 8053063681,
-                    8589934593, 9126805505, 9663676417, 10200547329, 10737418241]
-      }
+      {
+        size: OBJ_SIZES.gb10p2,
+        sourceConfig: new CopySourceOptions({ Start: -1 }),
+        expectedStart: [
+          0, 536870913, 1073741826, 1610612738, 2147483650, 2684354562, 3221225474, 3758096386, 4294967298, 4831838210,
+          5368709122, 5905580034, 6442450946, 6979321858, 7516192770, 8053063682, 8589934594, 9126805506, 9663676418,
+          10200547330,
+        ],
+        expectedEnd: [
+          536870912, 1073741825, 1610612737, 2147483649, 2684354561, 3221225473, 3758096385, 4294967297, 4831838209,
+          5368709121, 5905580033, 6442450945, 6979321857, 7516192769, 8053063681, 8589934593, 9126805505, 9663676417,
+          10200547329, 10737418241,
+        ],
+      },
     ]
 
-
-    expectedSplitsTestCases.forEach((testCase)=>{
+    expectedSplitsTestCases.forEach((testCase) => {
       const fnResult = calculateEvenSplits(testCase.size, testCase)
-      const {
-        startIndex,
-        endIndex
-      } = fnResult ||{}
+      const { startIndex, endIndex } = fnResult || {}
 
-      if(isArray(startIndex) && isArray(endIndex)) {
-        const isExpectedResult = (startIndex.length === testCase.expectedStart.length) && (endIndex.length === testCase.expectedEnd.length)
+      if (isArray(startIndex) && isArray(endIndex)) {
+        const isExpectedResult =
+          startIndex.length === testCase.expectedStart.length && endIndex.length === testCase.expectedEnd.length
         assert.equal(isExpectedResult, true)
-      } else{ // null cases.
+      } else {
+        // null cases.
         assert.equal(startIndex, expectedSplitsTestCases.expectedStart)
         assert.equal(endIndex, expectedSplitsTestCases.expectedEnd)
       }
     })
-      
   })
-    
 })
 
 describe('CopyConditions', () => {
@@ -204,7 +230,7 @@ describe('CopyConditions', () => {
   })
 })
 
-describe('Client', function() {
+describe('Client', function () {
   var nockRequests = []
   this.timeout(5000)
   beforeEach(() => {
@@ -212,7 +238,7 @@ describe('Client', function() {
     nockRequests = []
   })
   afterEach(() => {
-    nockRequests.forEach(element => {
+    nockRequests.forEach((element) => {
       if (!element.request.isDone()) {
         element.request.done()
       }
@@ -223,14 +249,14 @@ describe('Client', function() {
     port: 9000,
     accessKey: 'accesskey',
     secretKey: 'secretkey',
-    useSSL: false
+    useSSL: false,
   })
   describe('new client', () => {
     it('should work with https', () => {
       var client = new Minio.Client({
         endPoint: 'localhost',
         accessKey: 'accesskey',
-        secretKey: 'secretkey'
+        secretKey: 'secretkey',
       })
       assert.equal(client.port, 443)
     })
@@ -240,7 +266,7 @@ describe('Client', function() {
         port: 9000,
         accessKey: 'accesskey',
         secretKey: 'secretkey',
-        useSSL: false
+        useSSL: false,
       })
       assert.equal(client.port, 9000)
     })
@@ -249,7 +275,7 @@ describe('Client', function() {
         endPoint: 'localhost',
         accessKey: 'accesskey',
         secretKey: 'secretkey',
-        useSSL: false
+        useSSL: false,
       })
       assert.equal(client.port, 80)
     })
@@ -258,7 +284,7 @@ describe('Client', function() {
         endPoint: 'localhost',
         port: 9000,
         accessKey: 'accesskey',
-        secretKey: 'secretkey'
+        secretKey: 'secretkey',
       })
       assert.equal(client.port, 9000)
     })
@@ -267,7 +293,7 @@ describe('Client', function() {
         new Minio.Client({
           endPoint: 'http://localhost:9000',
           accessKey: 'accesskey',
-          secretKey: 'secretkey'
+          secretKey: 'secretkey',
         })
       } catch (e) {
         done()
@@ -278,7 +304,7 @@ describe('Client', function() {
         new Minio.Client({
           endPoint: 'localhost##$@3',
           accessKey: 'accesskey',
-          secretKey: 'secretkey'
+          secretKey: 'secretkey',
         })
       } catch (e) {
         done()
@@ -288,7 +314,7 @@ describe('Client', function() {
       try {
         new Minio.Client({
           accessKey: 'accesskey',
-          secretKey: 'secretkey'
+          secretKey: 'secretkey',
         })
       } catch (e) {
         done()
@@ -300,7 +326,7 @@ describe('Client', function() {
           endPoint: 'localhost',
           port: -1,
           accessKey: 'accesskey',
-          secretKey: 'secretkey'
+          secretKey: 'secretkey',
         })
       } catch (e) {
         done()
@@ -313,7 +339,7 @@ describe('Client', function() {
           secure: false,
           port: 9000,
           accessKey: 'accesskey',
-          secretKey: 'secretkey'
+          secretKey: 'secretkey',
         })
       } catch (e) {
         done()
@@ -326,7 +352,7 @@ describe('Client', function() {
           secure: true,
           port: 9000,
           accessKey: 'accesskey',
-          secretKey: 'secretkey'
+          secretKey: 'secretkey',
         })
       } catch (e) {
         done()
@@ -340,16 +366,16 @@ describe('Client', function() {
           var client = new Minio.Client({
             endPoint: 'localhost',
             port: 9000,
-            useSSL: false
+            useSSL: false,
           })
-          client.presignedGetObject('bucket', 'object', 1000, function() {})
+          client.presignedGetObject('bucket', 'object', 1000, function () {})
         } catch (e) {
           done()
         }
       })
       it('should not generate presigned url with wrong expires param', (done) => {
         try {
-          client.presignedGetObject('bucket', 'object', '0', function() {})
+          client.presignedGetObject('bucket', 'object', '0', function () {})
         } catch (e) {
           done()
         }
@@ -361,16 +387,16 @@ describe('Client', function() {
           var client = new Minio.Client({
             endPoint: 'localhost',
             port: 9000,
-            useSSL: false
+            useSSL: false,
           })
-          client.presignedPutObject('bucket', 'object', 1000, function() {})
+          client.presignedPutObject('bucket', 'object', 1000, function () {})
         } catch (e) {
           done()
         }
       })
       it('should not generate presigned url with wrong expires param', (done) => {
         try {
-          client.presignedPutObject('bucket', 'object', '0', function() {})
+          client.presignedPutObject('bucket', 'object', '0', function () {})
         } catch (e) {
           done()
         }
@@ -414,37 +440,40 @@ describe('Client', function() {
       var client = new Minio.Client({
         endPoint: 'localhost',
         accessKey: 'accesskey',
-        secretKey: 'secretkey'
+        secretKey: 'secretkey',
       })
-      assert.equal(`MinIO (${process.platform}; ${process.arch}) minio-js/${Package.version}`,
-                   client.userAgent)
+      assert.equal(`MinIO (${process.platform}; ${process.arch}) minio-js/${Package.version}`, client.userAgent)
     })
     it('should set user agent', () => {
       var client = new Minio.Client({
         endPoint: 'localhost',
         accessKey: 'accesskey',
-        secretKey: 'secretkey'
+        secretKey: 'secretkey',
       })
       client.setAppInfo('test', '3.2.1')
-      assert.equal(`MinIO (${process.platform}; ${process.arch}) minio-js/${Package.version} test/3.2.1`,
-                   client.userAgent)
+      assert.equal(
+        `MinIO (${process.platform}; ${process.arch}) minio-js/${Package.version} test/3.2.1`,
+        client.userAgent,
+      )
     })
     it('should set user agent without comments', () => {
       var client = new Minio.Client({
         endPoint: 'localhost',
         accessKey: 'accesskey',
-        secretKey: 'secretkey'
+        secretKey: 'secretkey',
       })
       client.setAppInfo('test', '3.2.1')
-      assert.equal(`MinIO (${process.platform}; ${process.arch}) minio-js/${Package.version} test/3.2.1`,
-                   client.userAgent)
+      assert.equal(
+        `MinIO (${process.platform}; ${process.arch}) minio-js/${Package.version} test/3.2.1`,
+        client.userAgent,
+      )
     })
     it('should not set user agent without name', (done) => {
       try {
         var client = new Minio.Client({
           endPoint: 'localhost',
           accessKey: 'accesskey',
-          secretKey: 'secretkey'
+          secretKey: 'secretkey',
         })
         client.setAppInfo(null, '3.2.1')
       } catch (e) {
@@ -456,7 +485,7 @@ describe('Client', function() {
         var client = new Minio.Client({
           endPoint: 'localhost',
           accessKey: 'accesskey',
-          secretKey: 'secretkey'
+          secretKey: 'secretkey',
         })
         client.setAppInfo('', '3.2.1')
       } catch (e) {
@@ -468,7 +497,7 @@ describe('Client', function() {
         var client = new Minio.Client({
           endPoint: 'localhost',
           accessKey: 'accesskey',
-          secretKey: 'secretkey'
+          secretKey: 'secretkey',
         })
         client.setAppInfo('test', null)
       } catch (e) {
@@ -480,7 +509,7 @@ describe('Client', function() {
         var client = new Minio.Client({
           endPoint: 'localhost',
           accessKey: 'accesskey',
-          secretKey: 'secretkey'
+          secretKey: 'secretkey',
         })
         client.setAppInfo('test', '')
       } catch (e) {
@@ -493,35 +522,35 @@ describe('Client', function() {
     describe('#getObject(bucket, object, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.getObject(null, 'hello', function() {})
+          client.getObject(null, 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.getObject('', 'hello', function() {})
+          client.getObject('', 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.getObject('  \n  \t  ', 'hello', function() {})
+          client.getObject('  \n  \t  ', 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on null object', (done) => {
         try {
-          client.getObject('hello', null, function() {})
+          client.getObject('hello', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty object', (done) => {
         try {
-          client.getObject('hello', '', function() {})
+          client.getObject('hello', '', function () {})
         } catch (e) {
           done()
         }
@@ -532,7 +561,7 @@ describe('Client', function() {
       describe('with small objects using single put', () => {
         it('should fail when data is smaller than specified', (done) => {
           var s = new Stream.Readable()
-          s._read = function() {}
+          s._read = function () {}
           s.push('hello world')
           s.push(null)
           client.putObject('bucket', 'object', s, 12, '', (e) => {
@@ -543,7 +572,7 @@ describe('Client', function() {
         })
         it('should fail when data is larger than specified', (done) => {
           var s = new Stream.Readable()
-          s._read = function() {}
+          s._read = function () {}
           s.push('hello world')
           s.push(null)
           client.putObject('bucket', 'object', s, 10, '', (e) => {
@@ -569,35 +598,35 @@ describe('Client', function() {
         })
         it('should fail on null bucket', (done) => {
           try {
-            client.putObject(null, 'hello', null, 1, '', function() {})
+            client.putObject(null, 'hello', null, 1, '', function () {})
           } catch (e) {
             done()
           }
         })
         it('should fail on empty bucket', (done) => {
           try {
-            client.putObject(' \n \t ', 'hello', null, 1, '', function() {})
+            client.putObject(' \n \t ', 'hello', null, 1, '', function () {})
           } catch (e) {
             done()
           }
         })
         it('should fail on empty bucket', (done) => {
           try {
-            client.putObject('', 'hello', null, 1, '', function() {})
+            client.putObject('', 'hello', null, 1, '', function () {})
           } catch (e) {
             done()
           }
         })
         it('should fail on null object', (done) => {
           try {
-            client.putObject('hello', null, null, 1, '', function() {})
+            client.putObject('hello', null, null, 1, '', function () {})
           } catch (e) {
             done()
           }
         })
         it('should fail on empty object', (done) => {
           try {
-            client.putObject('hello', '', null, 1, '', function() {})
+            client.putObject('hello', '', null, 1, '', function () {})
           } catch (e) {
             done()
           }
@@ -608,7 +637,11 @@ describe('Client', function() {
     describe('#removeAllBucketNotification()', () => {
       it('should error on invalid arguments', () => {
         assert.throws(() => {
-          client.removeAllBucketNotification('ab', () => {}, function() {})
+          client.removeAllBucketNotification(
+            'ab',
+            () => {},
+            function () {},
+          )
         }, /Invalid bucket name/)
       })
     })
@@ -652,35 +685,35 @@ describe('Client', function() {
     describe('#statObject(bucket, object, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.statObject(null, 'hello', function() {})
+          client.statObject(null, 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.statObject('', 'hello', function() {})
+          client.statObject('', 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.statObject('  \n  \t  ', 'hello', function() {})
+          client.statObject('  \n  \t  ', 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on null object', (done) => {
         try {
-          client.statObject('hello', null, function() {})
+          client.statObject('hello', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty object', (done) => {
         try {
-          client.statObject('hello', '', function() {})
+          client.statObject('hello', '', function () {})
         } catch (e) {
           done()
         }
@@ -688,60 +721,59 @@ describe('Client', function() {
 
       it('should fail on incompatible argument type (number) for statOpts object', (done) => {
         try {
-          client.statObject('hello', 'testStatOpts', 1, function() {})
+          client.statObject('hello', 'testStatOpts', 1, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on incompatible argument type (null) for statOpts object', (done) => {
         try {
-          client.statObject('hello', 'testStatOpts', null, function() {})
+          client.statObject('hello', 'testStatOpts', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on incompatible argument type (sting) for statOpts object', (done) => {
         try {
-          client.statObject('hello', 'testStatOpts', '  ', function() {})
+          client.statObject('hello', 'testStatOpts', '  ', function () {})
         } catch (e) {
           done()
         }
       })
-        
     })
 
     describe('#removeObject(bucket, object, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.removeObject(null, 'hello', function() {})
+          client.removeObject(null, 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.removeObject('', 'hello', function() {})
+          client.removeObject('', 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.removeObject('  \n  \t  ', 'hello', function() {})
+          client.removeObject('  \n  \t  ', 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on null object', (done) => {
         try {
-          client.removeObject('hello', null, function() {})
+          client.removeObject('hello', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty object', (done) => {
         try {
-          client.removeObject('hello', '', function() {})
+          client.removeObject('hello', '', function () {})
         } catch (e) {
           done()
         }
@@ -749,15 +781,15 @@ describe('Client', function() {
       // Versioning related options as removeOpts
       it('should fail on empty (null) removeOpts object', (done) => {
         try {
-          client.removeObject('hello', 'testRemoveOpts',null, function() {})
+          client.removeObject('hello', 'testRemoveOpts', null, function () {})
         } catch (e) {
           done()
         }
       })
-      
+
       it('should fail on empty (string) removeOpts', (done) => {
         try {
-          client.removeObject('hello', 'testRemoveOpts','', function() {})
+          client.removeObject('hello', 'testRemoveOpts', '', function () {})
         } catch (e) {
           done()
         }
@@ -767,35 +799,35 @@ describe('Client', function() {
     describe('#removeIncompleteUpload(bucket, object, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.removeIncompleteUpload(null, 'hello', function() {})
+          client.removeIncompleteUpload(null, 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.removeIncompleteUpload('', 'hello', function() {})
+          client.removeIncompleteUpload('', 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.removeIncompleteUpload('  \n  \t  ', 'hello', function() {})
+          client.removeIncompleteUpload('  \n  \t  ', 'hello', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on null object', (done) => {
         try {
-          client.removeIncompleteUpload('hello', null, function() {})
+          client.removeIncompleteUpload('hello', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty object', (done) => {
         try {
-          client.removeIncompleteUpload('hello', '', function() {})
+          client.removeIncompleteUpload('hello', '', function () {})
         } catch (e) {
           done()
         }
@@ -803,20 +835,18 @@ describe('Client', function() {
     })
   })
 
-  describe('Bucket Versioning APIs', ()=>{
+  describe('Bucket Versioning APIs', () => {
     describe('getBucketVersioning(bucket, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.getBucketVersioning(null, function () {
-          })
+          client.getBucketVersioning(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.getBucketVersioning('', function () {
-          })
+          client.getBucketVersioning('', function () {})
         } catch (e) {
           done()
         }
@@ -826,16 +856,14 @@ describe('Client', function() {
     describe('setBucketVersioning(bucket, versionConfig, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.setBucketVersioning(null, {},function () {
-          })
+          client.setBucketVersioning(null, {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.setBucketVersioning('', {},function () {
-          })
+          client.setBucketVersioning('', {}, function () {})
         } catch (e) {
           done()
         }
@@ -843,8 +871,7 @@ describe('Client', function() {
 
       it('should fail on empty versionConfig', (done) => {
         try {
-          client.setBucketVersioning('', null,function () {
-          })
+          client.setBucketVersioning('', null, function () {})
         } catch (e) {
           done()
         }
@@ -852,88 +879,75 @@ describe('Client', function() {
     })
   })
 
-
   describe('Bucket and Object Tags APIs', () => {
     describe('Set Bucket Tags ', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.setBucketTagging(null, {}, function () {
-          })
+          client.setBucketTagging(null, {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.setBucketTagging('', {}, function () {
-          })
+          client.setBucketTagging('', {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail if tags are more than 50', (done) => {
-        const _50_plus_key_tags={}
-        for(let i=0;i<51;i+=1){
-          _50_plus_key_tags[i]=i
+        const _50_plus_key_tags = {}
+        for (let i = 0; i < 51; i += 1) {
+          _50_plus_key_tags[i] = i
         }
         try {
-          client.setBucketTagging('', _50_plus_key_tags, function () {
-          })
+          client.setBucketTagging('', _50_plus_key_tags, function () {})
         } catch (e) {
           done()
         }
       })
-
     })
     describe('Get Bucket Tags', () => {
       it('should fail on invalid bucket', (done) => {
         try {
-          client.getBucketTagging('nv',null,  function () {
-          })
+          client.getBucketTagging('nv', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on null bucket', (done) => {
         try {
-          client.getBucketTagging(null, function () {
-          })
+          client.getBucketTagging(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.getBucketTagging('', function () {
-          })
+          client.getBucketTagging('', function () {})
         } catch (e) {
           done()
         }
       })
-
-
     })
     describe('Remove Bucket Tags', () => {
       it('should fail on null object', (done) => {
         try {
-          client.removeBucketTagging(null, function () {
-          })
+          client.removeBucketTagging(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.removeBucketTagging('',  function () {
-          })
+          client.removeBucketTagging('', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on invalid bucket name', (done) => {
         try {
-          client.removeBucketTagging('198.51.100.24', function () {
-          })
+          client.removeBucketTagging('198.51.100.24', function () {})
         } catch (e) {
           done()
         }
@@ -941,8 +955,7 @@ describe('Client', function() {
 
       it('should fail on invalid bucket name', (done) => {
         try {
-          client.removeBucketTagging('xy', function () {
-          })
+          client.removeBucketTagging('xy', function () {})
         } catch (e) {
           done()
         }
@@ -951,91 +964,78 @@ describe('Client', function() {
     describe('Put Object Tags', () => {
       it('should fail on null object', (done) => {
         try {
-          client.putObjectTagging('my-bucket-name',null, {}, function () {
-          })
+          client.putObjectTagging('my-bucket-name', null, {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty object', (done) => {
         try {
-          client.putObjectTagging('my-bucket-name',null, {}, function () {
-          })
+          client.putObjectTagging('my-bucket-name', null, {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on non object tags', (done) => {
         try {
-          client.putObjectTagging('my-bucket-name',null, 'non-obj-tag', function () {
-          })
+          client.putObjectTagging('my-bucket-name', null, 'non-obj-tag', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail if tags are more than 50 on an object', (done) => {
-        const _50_plus_key_tags={}
-        for(let i=0;i<51;i+=1){
-          _50_plus_key_tags[i]=i
+        const _50_plus_key_tags = {}
+        for (let i = 0; i < 51; i += 1) {
+          _50_plus_key_tags[i] = i
         }
         try {
-          client.putObjectTagging('my-bucket-name',null, _50_plus_key_tags, function () {
-          })
+          client.putObjectTagging('my-bucket-name', null, _50_plus_key_tags, function () {})
         } catch (e) {
           done()
         }
       })
-
     })
     describe('Get Object Tags', () => {
       it('should fail on invalid bucket', (done) => {
         try {
-          client.getObjectTagging('nv',null,  function () {
-          })
+          client.getObjectTagging('nv', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on null object', (done) => {
         try {
-          client.getObjectTagging('my-bucket-name',null, function () {
-          })
+          client.getObjectTagging('my-bucket-name', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty object', (done) => {
         try {
-          client.getObjectTagging('my-bucket-name',null, function () {
-          })
+          client.getObjectTagging('my-bucket-name', null, function () {})
         } catch (e) {
           done()
         }
       })
-
-
     })
     describe('Remove Object Tags', () => {
       it('should fail on null object', (done) => {
         try {
-          client.removeObjectTagging('my-bucket',null, function () {
-          })
+          client.removeObjectTagging('my-bucket', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.removeObjectTagging('my-bucket', '',  function () {
-          })
+          client.removeObjectTagging('my-bucket', '', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on invalid bucket name', (done) => {
         try {
-          client.removeObjectTagging('198.51.100.24', function () {
-          })
+          client.removeObjectTagging('198.51.100.24', function () {})
         } catch (e) {
           done()
         }
@@ -1043,8 +1043,7 @@ describe('Client', function() {
 
       it('should fail on invalid bucket name', (done) => {
         try {
-          client.removeObjectTagging('xy', function () {
-          })
+          client.removeObjectTagging('xy', function () {})
         } catch (e) {
           done()
         }
@@ -1052,12 +1051,10 @@ describe('Client', function() {
     })
   })
 
-
   describe('setBucketLifecycle(bucket, lifecycleConfig, callback)', () => {
     it('should fail on null bucket', (done) => {
       try {
-        client.setBucketLifecycle(null, null,function () {
-        })
+        client.setBucketLifecycle(null, null, function () {})
       } catch (e) {
         done()
       }
@@ -1065,20 +1062,17 @@ describe('Client', function() {
 
     it('should fail on empty bucket', (done) => {
       try {
-        client.setBucketLifecycle('', null,function () {
-        })
+        client.setBucketLifecycle('', null, function () {})
       } catch (e) {
         done()
       }
     })
-
   })
 
   describe('getBucketLifecycle(bucket, callback)', () => {
     it('should fail on null bucket', (done) => {
       try {
-        client.getBucketLifecycle(null, function () {
-        })
+        client.getBucketLifecycle(null, function () {})
       } catch (e) {
         done()
       }
@@ -1086,19 +1080,16 @@ describe('Client', function() {
 
     it('should fail on empty bucket', (done) => {
       try {
-        client.getBucketLifecycle('', function () {
-        })
+        client.getBucketLifecycle('', function () {})
       } catch (e) {
         done()
       }
     })
-
   })
   describe('removeBucketLifecycle(bucket, callback)', () => {
     it('should fail on null bucket', (done) => {
       try {
-        client.removeBucketLifecycle(null, null,function () {
-        })
+        client.removeBucketLifecycle(null, null, function () {})
       } catch (e) {
         done()
       }
@@ -1106,29 +1097,25 @@ describe('Client', function() {
 
     it('should fail on empty bucket', (done) => {
       try {
-        client.removeBucketLifecycle('', null,function () {
-        })
+        client.removeBucketLifecycle('', null, function () {})
       } catch (e) {
         done()
       }
     })
-
   })
 
-  describe('Object Locking APIs', ()=> {
+  describe('Object Locking APIs', () => {
     describe('getObjectLockConfig(bucket, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.getObjectLockConfig(null, function () {
-          })
+          client.getObjectLockConfig(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.getObjectLockConfig('', function () {
-          })
+          client.getObjectLockConfig('', function () {})
         } catch (e) {
           done()
         }
@@ -1138,48 +1125,50 @@ describe('Client', function() {
     describe('setObjectLockConfig(bucket, lockConfig, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.setObjectLockConfig(null, function () {
-          })
+          client.setObjectLockConfig(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.setObjectLockConfig('', function () {
-          })
+          client.setObjectLockConfig('', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on passing invalid mode ', (done) => {
         try {
-          client.setObjectLockConfig('my-bucket',{mode:"invalid_mode"}, function () {
-          })
+          client.setObjectLockConfig('my-bucket', { mode: 'invalid_mode' }, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on passing invalid unit ', (done) => {
         try {
-          client.setObjectLockConfig('my-bucket',{ mode:"COMPLIANCE",unit:"invalid_unit"}, function () {
-          })
+          client.setObjectLockConfig('my-bucket', { mode: 'COMPLIANCE', unit: 'invalid_unit' }, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on passing invalid validity ', (done) => {
         try {
-          client.setObjectLockConfig('my-bucket',{mode:"COMPLIANCE",unit:"invalid_unit", validity:''}, function () {
-          })
+          client.setObjectLockConfig(
+            'my-bucket',
+            { mode: 'COMPLIANCE', unit: 'invalid_unit', validity: '' },
+            function () {},
+          )
         } catch (e) {
           done()
         }
       })
       it('should fail on passing  invalid config ', (done) => {
         try {
-          client.setObjectLockConfig('my-bucket',{mode:"COMPLIANCE", randomProp:true, nonExisting:false,}, function () {
-          })
+          client.setObjectLockConfig(
+            'my-bucket',
+            { mode: 'COMPLIANCE', randomProp: true, nonExisting: false },
+            function () {},
+          )
         } catch (e) {
           done()
         }
@@ -1187,36 +1176,32 @@ describe('Client', function() {
     })
   })
 
-  describe('Object retention APIs', ()=> {
+  describe('Object retention APIs', () => {
     describe('getObjectRetention(bucket, objectName, getRetentionOpts,callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.getObjectRetention(null,'','', function () {
-          })
+          client.getObjectRetention(null, '', '', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.getObjectRetention('', '','',function () {
-          })
+          client.getObjectRetention('', '', '', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on invalid  object name', (done) => {
         try {
-          client.getObjectRetention('my-bucket', null, '',function () {
-          })
+          client.getObjectRetention('my-bucket', null, '', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on invalid  versionId', (done) => {
         try {
-          client.getObjectRetention('my-bucket', 'objectname', {versionId:123},function () {
-          })
+          client.getObjectRetention('my-bucket', 'objectname', { versionId: 123 }, function () {})
         } catch (e) {
           done()
         }
@@ -1226,16 +1211,14 @@ describe('Client', function() {
     describe('putObjectRetention(bucket, objectName, retentionConfig, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.putObjectRetention(null,'',{}, function () {
-          })
+          client.putObjectRetention(null, '', {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.putObjectRetention('','',{}, function () {
-          })
+          client.putObjectRetention('', '', {}, function () {})
         } catch (e) {
           done()
         }
@@ -1243,48 +1226,42 @@ describe('Client', function() {
 
       it('should fail on null object', (done) => {
         try {
-          client.putObjectRetention('my-bucket',null,{}, function () {
-          })
+          client.putObjectRetention('my-bucket', null, {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty object', (done) => {
         try {
-          client.putObjectRetention('my-bucket', '', {}, function () {
-          })
+          client.putObjectRetention('my-bucket', '', {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on passing invalid mode ', (done) => {
         try {
-          client.putObjectRetention('my-bucket', 'my-object', {mode:"invalid_mode"}, function () {
-          })
+          client.putObjectRetention('my-bucket', 'my-object', { mode: 'invalid_mode' }, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on passing invalid governanceBypass ', (done) => {
         try {
-          client.putObjectRetention('my-bucket', 'my-object', { governanceBypass:"nonbool"}, function () {
-          })
+          client.putObjectRetention('my-bucket', 'my-object', { governanceBypass: 'nonbool' }, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on passing invalid (null) retainUntilDate ', (done) => {
         try {
-          client.putObjectRetention('my-bucket', 'my-object', { retainUntilDate:12345}, function () {
-          })
+          client.putObjectRetention('my-bucket', 'my-object', { retainUntilDate: 12345 }, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on passing invalid versionId ', (done) => {
         try {
-          client.putObjectRetention('my-bucket',{ versionId:"COMPLIANCE" }, function () {
-          })
+          client.putObjectRetention('my-bucket', { versionId: 'COMPLIANCE' }, function () {})
         } catch (e) {
           done()
         }
@@ -1292,65 +1269,60 @@ describe('Client', function() {
     })
   })
 
-
-  describe('Bucket Encryption APIs', ()=> {
-
+  describe('Bucket Encryption APIs', () => {
     describe('setBucketEncryption(bucket, encryptionConfig, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.setBucketEncryption(null, function () {
-          })
+          client.setBucketEncryption(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.setBucketEncryption('', function () {
-          })
+          client.setBucketEncryption('', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on multiple rules', (done) => {
         try {
-          client.setBucketEncryption('my-bucket', {
-            // Default Rule
-            Rule:[
-              {
-                ApplyServerSideEncryptionByDefault: {
-                  SSEAlgorithm:"AES256"
-                }
-              },
-              {
-                ApplyServerSideEncryptionByDefault: {
-                  SSEAlgorithm:"AES256"
-                }
-              }
-            ]
-
-          },function () {
-          })
+          client.setBucketEncryption(
+            'my-bucket',
+            {
+              // Default Rule
+              Rule: [
+                {
+                  ApplyServerSideEncryptionByDefault: {
+                    SSEAlgorithm: 'AES256',
+                  },
+                },
+                {
+                  ApplyServerSideEncryptionByDefault: {
+                    SSEAlgorithm: 'AES256',
+                  },
+                },
+              ],
+            },
+            function () {},
+          )
         } catch (e) {
           done()
         }
       })
-
     })
 
     describe('getBucketEncryption(bucket, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.getBucketEncryption(null, function () {
-          })
+          client.getBucketEncryption(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.getBucketEncryption('', function () {
-          })
+          client.getBucketEncryption('', function () {})
         } catch (e) {
           done()
         }
@@ -1360,37 +1332,32 @@ describe('Client', function() {
     describe('removeBucketEncryption(bucket, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.removeBucketEncryption(null, function () {
-          })
+          client.removeBucketEncryption(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.removeBucketEncryption('', function () {
-          })
+          client.removeBucketEncryption('', function () {})
         } catch (e) {
           done()
         }
       })
     })
-
   })
-  describe('Bucket Replication APIs', ()=> {
+  describe('Bucket Replication APIs', () => {
     describe('setBucketReplication(bucketName, replicationConfig, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.setBucketReplication(null, {}, function () {
-          })
+          client.setBucketReplication(null, {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.setBucketReplication('', {}, function () {
-          })
+          client.setBucketReplication('', {}, function () {})
         } catch (e) {
           done()
         }
@@ -1398,8 +1365,7 @@ describe('Client', function() {
 
       it('should fail on empty replicationConfig', (done) => {
         try {
-          client.setBucketReplication('my-bucket', {}, function () {
-          })
+          client.setBucketReplication('my-bucket', {}, function () {})
         } catch (e) {
           done()
         }
@@ -1407,8 +1373,7 @@ describe('Client', function() {
 
       it('should fail on empty replicationConfig role', (done) => {
         try {
-          client.setBucketReplication('my-bucket', {role:''}, function () {
-          })
+          client.setBucketReplication('my-bucket', { role: '' }, function () {})
         } catch (e) {
           done()
         }
@@ -1416,8 +1381,7 @@ describe('Client', function() {
 
       it('should fail on  invalid value for replicationConfig role', (done) => {
         try {
-          client.setBucketReplication('my-bucket', {role:12}, function () {
-          })
+          client.setBucketReplication('my-bucket', { role: 12 }, function () {})
         } catch (e) {
           done()
         }
@@ -1425,36 +1389,31 @@ describe('Client', function() {
 
       it('should fail on  empty value for replicationConfig rules', (done) => {
         try {
-          client.setBucketReplication('my-bucket', {role:"arn:",rules:[]}, function () {
-          })
+          client.setBucketReplication('my-bucket', { role: 'arn:', rules: [] }, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on  null value for replicationConfig rules', (done) => {
         try {
-          client.setBucketReplication('my-bucket', {role:"arn:",rules:null}, function () {
-          })
+          client.setBucketReplication('my-bucket', { role: 'arn:', rules: null }, function () {})
         } catch (e) {
           done()
         }
       })
-
     })
 
     describe('getBucketReplication(bucketName, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.getBucketReplication(null, {}, function () {
-          })
+          client.getBucketReplication(null, {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.getBucketReplication('', {},function () {
-          })
+          client.getBucketReplication('', {}, function () {})
         } catch (e) {
           done()
         }
@@ -1464,16 +1423,14 @@ describe('Client', function() {
     describe('removeBucketReplication(bucketName, callback)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.removeBucketReplication(null, {}, function () {
-          })
+          client.removeBucketReplication(null, {}, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.removeBucketReplication('', {},function () {
-          })
+          client.removeBucketReplication('', {}, function () {})
         } catch (e) {
           done()
         }
@@ -1481,21 +1438,18 @@ describe('Client', function() {
     })
   })
 
-
-  describe('Object Legal Hold APIs', ()=> {
+  describe('Object Legal Hold APIs', () => {
     describe('getObjectLegalHold(bucketName, objectName, getOpts={}, cb)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.getObjectLegalHold(null, function () {
-          })
+          client.getObjectLegalHold(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.getObjectLegalHold('', function () {
-          })
+          client.getObjectLegalHold('', function () {})
         } catch (e) {
           done()
         }
@@ -1503,16 +1457,14 @@ describe('Client', function() {
 
       it('should fail on null objectName', (done) => {
         try {
-          client.getObjectLegalHold('my-bucket', null, function () {
-          })
+          client.getObjectLegalHold('my-bucket', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on null getOpts', (done) => {
         try {
-          client.getObjectLegalHold('my-bucker', 'my-object', null,  function () {
-          })
+          client.getObjectLegalHold('my-bucker', 'my-object', null, function () {})
         } catch (e) {
           done()
         }
@@ -1522,16 +1474,14 @@ describe('Client', function() {
     describe('setObjectLegalHold(bucketName, objectName, setOpts={}, cb)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.setObjectLegalHold(null, function () {
-          })
+          client.setObjectLegalHold(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.setObjectLegalHold('', function () {
-          })
+          client.setObjectLegalHold('', function () {})
         } catch (e) {
           done()
         }
@@ -1539,24 +1489,21 @@ describe('Client', function() {
 
       it('should fail on null objectName', (done) => {
         try {
-          client.setObjectLegalHold('my-bucket', null, function () {
-          })
+          client.setObjectLegalHold('my-bucket', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on null setOpts', (done) => {
         try {
-          client.setObjectLegalHold('my-bucker', 'my-object', null,  function () {
-          })
+          client.setObjectLegalHold('my-bucker', 'my-object', null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty versionId', (done) => {
         try {
-          client.setObjectLegalHold('my-bucker', 'my-object', {}, function () {
-          })
+          client.setObjectLegalHold('my-bucker', 'my-object', {}, function () {})
         } catch (e) {
           done()
         }
@@ -1564,12 +1511,11 @@ describe('Client', function() {
     })
   })
 
-  describe('Compose Object APIs', ()=> {
+  describe('Compose Object APIs', () => {
     describe('composeObject(destObjConfig, sourceObjectList,cb)', () => {
       it('should fail on null destination config', (done) => {
         try {
-          client.composeObject(null, function () {
-          })
+          client.composeObject(null, function () {})
         } catch (e) {
           done()
         }
@@ -1577,9 +1523,8 @@ describe('Client', function() {
 
       it('should fail on no array source config', (done) => {
         try {
-          const destOptions = new CopyDestinationOptions({Bucket:'test-bucket', Object:'test-object'})
-          client.composeObject(destOptions, 'non-array', function () {
-          })
+          const destOptions = new CopyDestinationOptions({ Bucket: 'test-bucket', Object: 'test-object' })
+          client.composeObject(destOptions, 'non-array', function () {})
         } catch (e) {
           done()
         }
@@ -1587,30 +1532,26 @@ describe('Client', function() {
 
       it('should fail on null source config', (done) => {
         try {
-          const destOptions = new CopyDestinationOptions({Bucket:'test-bucket', Object:'test-object'})
-          client.composeObject(destOptions, null, function () {
-          })
+          const destOptions = new CopyDestinationOptions({ Bucket: 'test-bucket', Object: 'test-object' })
+          client.composeObject(destOptions, null, function () {})
         } catch (e) {
           done()
         }
       })
-        
     })
   })
-  describe('Select Object Content APIs', ()=> {
+  describe('Select Object Content APIs', () => {
     describe('selectObjectContent(bucketName, objectName, selectOpts={}, cb)', () => {
       it('should fail on null bucket', (done) => {
         try {
-          client.selectObjectContent(null, function () {
-          })
+          client.selectObjectContent(null, function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on empty bucket', (done) => {
         try {
-          client.selectObjectContent('', function () {
-          })
+          client.selectObjectContent('', function () {})
         } catch (e) {
           done()
         }
@@ -1618,546 +1559,550 @@ describe('Client', function() {
 
       it('should fail on empty object', (done) => {
         try {
-          client.selectObjectContent('my-bucket','', function () {
-          })
+          client.selectObjectContent('my-bucket', '', function () {})
         } catch (e) {
           done()
         }
       })
       it('should fail on null object', (done) => {
         try {
-          client.selectObjectContent('my-bucket',null, function () {
-          })
+          client.selectObjectContent('my-bucket', null, function () {})
         } catch (e) {
           done()
         }
       })
-
     })
   })
-
 })
 
-
-describe('IP Address Validations', ()=>{
-
+describe('IP Address Validations', () => {
   it('should validate for valid ip', () => {
     assert.equal(isValidIP('1.1.1.1'), true)
   })
-    
+
   it('Check list of IPV4 Invalid addresses', () => {
-    const invalidIpv4 = [' 127.0.0.1', '127.0.0.1 ', '127.0.0.1 127.0.0.1', '127.0.0.256', '127.0.0.1//1', '127.0.0.1/0x1', '127.0.0.1/-1', '127.0.0.1/ab', '127.0.0.1/', '127.0.0.256/32', '127.0.0.1/33']
-    invalidIpv4.map(ip=>{
+    const invalidIpv4 = [
+      ' 127.0.0.1',
+      '127.0.0.1 ',
+      '127.0.0.1 127.0.0.1',
+      '127.0.0.256',
+      '127.0.0.1//1',
+      '127.0.0.1/0x1',
+      '127.0.0.1/-1',
+      '127.0.0.1/ab',
+      '127.0.0.1/',
+      '127.0.0.256/32',
+      '127.0.0.1/33',
+    ]
+    invalidIpv4.map((ip) => {
       assert.equal(isValidIP(ip), false)
     })
   })
 
   it('Check list of IPV4 Valid addresses', () => {
-    const validIpv4 =['001.002.003.004', '127.0.0.1', '255.255.255.255','192.168.1.10']
-    validIpv4.map(ip=>{
+    const validIpv4 = ['001.002.003.004', '127.0.0.1', '255.255.255.255', '192.168.1.10']
+    validIpv4.map((ip) => {
       assert.equal(isValidIP(ip), true)
     })
-
   })
 
   it('Check list of IPV6 Invalid addresses', () => {
-    const invalidIpV6 =[
+    const invalidIpV6 = [
       "':10.0.0.1",
-      "-1",
-      "::1 ::1",
-      "1.2.3.4:1111:2222:3333:4444::5555",
-      "1.2.3.4:1111:2222:3333::5555",
-      "1.2.3.4:1111:2222::5555",
-      "1.2.3.4:1111::5555",
-      "1.2.3.4::",
-      "1.2.3.4::5555",
-      "11112222:3333:4444:5555:6666:1.2.3.4",
-      "11112222:3333:4444:5555:6666:7777:8888",
-      "::1//64",
-      "::1/0001",
-      "1111:",
-      "1111:1.2.3.4",
-      "1111:2222",
-      "1111:22223333:4444:5555:6666:1.2.3.4",
-      "1111:22223333:4444:5555:6666:7777:8888",
-      "1111:2222:",
-      "1111:2222:1.2.3.4",
-      "1111:2222:3333",
-      "1111:2222:33334444:5555:6666:1.2.3.4",
-      "1111:2222:33334444:5555:6666:7777:8888",
-      "1111:2222:3333:",
-      "1111:2222:3333:1.2.3.4",
-      "1111:2222:3333:4444",
-      "1111:2222:3333:44445555:6666:1.2.3.4",
-      "1111:2222:3333:44445555:6666:7777:8888",
-      "1111:2222:3333:4444:",
-      "1111:2222:3333:4444:1.2.3.4",
-      "1111:2222:3333:4444:5555",
-      "1111:2222:3333:4444:55556666:1.2.3.4",
-      "1111:2222:3333:4444:55556666:7777:8888",
-      "1111:2222:3333:4444:5555:",
-      "1111:2222:3333:4444:5555:1.2.3.4",
-      "1111:2222:3333:4444:5555:6666",
-      "1111:2222:3333:4444:5555:66661.2.3.4",
-      "1111:2222:3333:4444:5555:66667777:8888",
-      "1111:2222:3333:4444:5555:6666:",
-      "1111:2222:3333:4444:5555:6666:1.2.3.4.5",
-      "1111:2222:3333:4444:5555:6666:255.255.255255",
-      "1111:2222:3333:4444:5555:6666:255.255255.255",
-      "1111:2222:3333:4444:5555:6666:255255.255.255",
-      "1111:2222:3333:4444:5555:6666:256.256.256.256",
-      "1111:2222:3333:4444:5555:6666:7777",
-      "1111:2222:3333:4444:5555:6666:77778888",
-      "1111:2222:3333:4444:5555:6666:7777:",
-      "1111:2222:3333:4444:5555:6666:7777:1.2.3.4",
-      "1111:2222:3333:4444:5555:6666:7777:::",
-      "1111:2222:3333:4444:5555:6666::8888:",
-      "1111:2222:3333:4444:5555:6666:::",
-      "1111:2222:3333:4444:5555:6666:::8888",
-      "1111:2222:3333:4444:5555::7777:8888:",
-      "1111:2222:3333:4444:5555::7777::",
-      "1111:2222:3333:4444:5555::8888:",
-      "1111:2222:3333:4444:5555:::",
-      "1111:2222:3333:4444:5555:::1.2.3.4",
-      "1111:2222:3333:4444:5555:::7777:8888",
-      "1111:2222:3333:4444::5555:",
-      "1111:2222:3333:4444::6666:7777:8888:",
-      "1111:2222:3333:4444::6666:7777::",
-      "1111:2222:3333:4444::6666::8888",
-      "1111:2222:3333:4444::7777:8888:",
-      "1111:2222:3333:4444::8888:",
-      "1111:2222:3333:4444:::",
-      "1111:2222:3333:4444:::6666:1.2.3.4",
-      "1111:2222:3333:4444:::6666:7777:8888",
-      "1111:2222:3333::5555:",
-      "1111:2222:3333::5555:6666:7777:8888:",
-      "1111:2222:3333::5555:6666:7777::",
-      "1111:2222:3333::5555:6666::8888",
-      "1111:2222:3333::5555::1.2.3.4",
-      "1111:2222:3333::5555::7777:8888",
-      "1111:2222:3333::6666:7777:8888:",
-      "1111:2222:3333::7777:8888:",
-      "1111:2222:3333::8888:",
-      "1111:2222:3333:::",
-      "1111:2222:3333:::5555:6666:1.2.3.4",
-      "1111:2222:3333:::5555:6666:7777:8888",
-      "1111:2222::4444:5555:6666:7777:8888:",
-      "1111:2222::4444:5555:6666:7777::",
-      "1111:2222::4444:5555:6666::8888",
-      "1111:2222::4444:5555::1.2.3.4",
-      "1111:2222::4444:5555::7777:8888",
-      "1111:2222::4444::6666:1.2.3.4",
-      "1111:2222::4444::6666:7777:8888",
-      "1111:2222::5555:",
-      "1111:2222::5555:6666:7777:8888:",
-      "1111:2222::6666:7777:8888:",
-      "1111:2222::7777:8888:",
-      "1111:2222::8888:",
-      "1111:2222:::",
-      "1111:2222:::4444:5555:6666:1.2.3.4",
-      "1111:2222:::4444:5555:6666:7777:8888",
-      "1111::3333:4444:5555:6666:7777:8888:",
-      "1111::3333:4444:5555:6666:7777::",
-      "1111::3333:4444:5555:6666::8888",
-      "1111::3333:4444:5555::1.2.3.4",
-      "1111::3333:4444:5555::7777:8888",
-      "1111::3333:4444::6666:1.2.3.4",
-      "1111::3333:4444::6666:7777:8888",
-      "1111::3333::5555:6666:1.2.3.4",
-      "1111::3333::5555:6666:7777:8888",
-      "1111::4444:5555:6666:7777:8888:",
-      "1111::5555:",
-      "1111::5555:6666:7777:8888:",
-      "1111::6666:7777:8888:",
-      "1111::7777:8888:",
-      "1111::8888:",
-      "1111:::",
-      "1111:::3333:4444:5555:6666:1.2.3.4",
-      "1111:::3333:4444:5555:6666:7777:8888",
-      "12345::6:7:8",
-      "124.15.6.89/60",
-      "1:2:3:4:5:6:7:8:9",
-      "1:2:3::4:5:6:7:8:9",
-      "1:2:3::4:5::7:8",
-      "1::1.2.256.4",
-      "1::1.2.3.256",
-      "1::1.2.3.300",
-      "1::1.2.3.900",
-      "1::1.2.300.4",
-      "1::1.2.900.4",
-      "1::1.256.3.4",
-      "1::1.300.3.4",
-      "1::1.900.3.4",
-      "1::256.2.3.4",
-      "1::260.2.3.4",
-      "1::2::3",
-      "1::300.2.3.4",
-      "1::300.300.300.300",
-      "1::3000.30.30.30",
-      "1::400.2.3.4",
-      "1::5:1.2.256.4",
-      "1::5:1.2.3.256",
-      "1::5:1.2.3.300",
-      "1::5:1.2.3.900",
-      "1::5:1.2.300.4",
-      "1::5:1.2.900.4",
-      "1::5:1.256.3.4",
-      "1::5:1.300.3.4",
-      "1::5:1.900.3.4",
-      "1::5:256.2.3.4",
-      "1::5:260.2.3.4",
-      "1::5:300.2.3.4",
-      "1::5:300.300.300.300",
-      "1::5:3000.30.30.30",
-      "1::5:400.2.3.4",
-      "1::5:900.2.3.4",
-      "1::900.2.3.4",
-      "1:::3:4:5",
-      "2001:0000:1234: 0000:0000:C1C0:ABCD:0876",
-      "2001:0000:1234:0000:0000:C1C0:ABCD:0876  0",
-      "2001:1:1:1:1:1:255Z255X255Y255",
-      "2001::FFD3::57ab",
-      "2001:DB8:0:0:8:800:200C:417A:221",
-      "2001:db8:85a3::8a2e:37023:7334",
-      "2001:db8:85a3::8a2e:370k:7334",
-      "3ffe:0b00:0000:0001:0000:0000:000a",
-      "3ffe:b00::1::a",
-      ":",
-      ":1.2.3.4",
-      ":1111:2222:3333:4444:5555:6666:1.2.3.4",
-      ":1111:2222:3333:4444:5555:6666:7777:8888",
-      ":1111:2222:3333:4444:5555:6666:7777::",
-      ":1111:2222:3333:4444:5555:6666::",
-      ":1111:2222:3333:4444:5555:6666::8888",
-      ":1111:2222:3333:4444:5555::",
-      ":1111:2222:3333:4444:5555::1.2.3.4",
-      ":1111:2222:3333:4444:5555::7777:8888",
-      ":1111:2222:3333:4444:5555::8888",
-      ":1111:2222:3333:4444::",
-      ":1111:2222:3333:4444::1.2.3.4",
-      ":1111:2222:3333:4444::5555",
-      ":1111:2222:3333:4444::6666:1.2.3.4",
-      ":1111:2222:3333:4444::6666:7777:8888",
-      ":1111:2222:3333:4444::7777:8888",
-      ":1111:2222:3333:4444::8888",
-      ":1111:2222:3333::",
-      ":1111:2222:3333::1.2.3.4",
-      ":1111:2222:3333::5555",
-      ":1111:2222:3333::5555:6666:1.2.3.4",
-      ":1111:2222:3333::5555:6666:7777:8888",
-      ":1111:2222:3333::6666:1.2.3.4",
-      ":1111:2222:3333::6666:7777:8888",
-      ":1111:2222:3333::7777:8888",
-      ":1111:2222:3333::8888",
-      ":1111:2222::",
-      ":1111:2222::1.2.3.4",
-      ":1111:2222::4444:5555:6666:1.2.3.4",
-      ":1111:2222::4444:5555:6666:7777:8888",
-      ":1111:2222::5555",
-      ":1111:2222::5555:6666:1.2.3.4",
-      ":1111:2222::5555:6666:7777:8888",
-      ":1111:2222::6666:1.2.3.4",
-      ":1111:2222::6666:7777:8888",
-      ":1111:2222::7777:8888",
-      ":1111:2222::8888",
-      ":1111::",
-      ":1111::1.2.3.4",
-      ":1111::3333:4444:5555:6666:1.2.3.4",
-      ":1111::3333:4444:5555:6666:7777:8888",
-      ":1111::4444:5555:6666:1.2.3.4",
-      ":1111::4444:5555:6666:7777:8888",
-      ":1111::5555",
-      ":1111::5555:6666:1.2.3.4",
-      ":1111::5555:6666:7777:8888",
-      ":1111::6666:1.2.3.4",
-      ":1111::6666:7777:8888",
-      ":1111::7777:8888",
-      ":1111::8888",
-      ":2222:3333:4444:5555:6666:1.2.3.4",
-      ":2222:3333:4444:5555:6666:7777:8888",
-      ":3333:4444:5555:6666:1.2.3.4",
-      ":3333:4444:5555:6666:7777:8888",
-      ":4444:5555:6666:1.2.3.4",
-      ":4444:5555:6666:7777:8888",
-      ":5555:6666:1.2.3.4",
-      ":5555:6666:7777:8888",
-      ":6666:1.2.3.4",
-      ":6666:7777:8888",
-      ":7777:8888",
-      ":8888",
-      "::-1",
-      "::.",
-      "::..",
-      "::...",
-      "::...4",
-      "::..3.",
-      "::..3.4",
-      "::.2..",
-      "::.2.3.",
-      "::.2.3.4",
-      "::1...",
-      "::1.2..",
-      "::1.2.256.4",
-      "::1.2.3.",
-      "::1.2.3.256",
-      "::1.2.3.300",
-      "::1.2.3.900",
-      "::1.2.300.4",
-      "::1.2.900.4",
-      "::1.256.3.4",
-      "::1.300.3.4",
-      "::1.900.3.4",
-      "::1111:2222:3333:4444:5555:6666::",
-      "::2222:3333:4444:5555:6666:7777:8888:",
-      "::2222:3333:4444:5555:7777:8888::",
-      "::2222:3333:4444:5555:7777::8888",
-      "::2222:3333:4444:5555::1.2.3.4",
-      "::2222:3333:4444:5555::7777:8888",
-      "::2222:3333:4444::6666:1.2.3.4",
-      "::2222:3333:4444::6666:7777:8888",
-      "::2222:3333::5555:6666:1.2.3.4",
-      "::2222:3333::5555:6666:7777:8888",
-      "::2222::4444:5555:6666:1.2.3.4",
-      "::2222::4444:5555:6666:7777:8888",
-      "::256.2.3.4",
-      "::260.2.3.4",
-      "::300.2.3.4",
-      "::300.300.300.300",
-      "::3000.30.30.30",
-      "::3333:4444:5555:6666:7777:8888:",
-      "::400.2.3.4",
-      "::4444:5555:6666:7777:8888:",
-      "::5555:",
-      "::5555:6666:7777:8888:",
-      "::6666:7777:8888:",
-      "::7777:8888:",
-      "::8888:",
-      "::900.2.3.4",
-      ":::",
-      ":::1.2.3.4",
-      ":::2222:3333:4444:5555:6666:1.2.3.4",
-      ":::2222:3333:4444:5555:6666:7777:8888",
-      ":::3333:4444:5555:6666:7777:8888",
-      ":::4444:5555:6666:1.2.3.4",
-      ":::4444:5555:6666:7777:8888",
-      ":::5555",
-      ":::5555:6666:1.2.3.4",
-      ":::5555:6666:7777:8888",
-      ":::6666:1.2.3.4",
-      ":::6666:7777:8888",
-      ":::7777:8888",
-      ":::8888",
-      "::ffff:192x168.1.26",
-      "::ffff:2.3.4",
-      "::ffff:257.1.2.3",
-      "FF01::101::2",
-      "FF02:0000:0000:0000:0000:0000:0000:0000:0001",
-      "XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:1.2.3.4",
-      "XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:XXXX",
-      "a::b::c",
-      "a::g",
-      "a:a:a:a:a:a:a:a:a",
-      "a:aaaaa::",
-      "a:b",
-      "a:b:c:d:e:f:g:0",
-      "ffff:",
-      "ffff::ffff::ffff",
-      "ffgg:ffff:ffff:ffff:ffff:ffff:ffff:ffff",
-      "ldkfj",
-      "::/129",
-      "1000:://32",
-      "::/"
+      '-1',
+      '::1 ::1',
+      '1.2.3.4:1111:2222:3333:4444::5555',
+      '1.2.3.4:1111:2222:3333::5555',
+      '1.2.3.4:1111:2222::5555',
+      '1.2.3.4:1111::5555',
+      '1.2.3.4::',
+      '1.2.3.4::5555',
+      '11112222:3333:4444:5555:6666:1.2.3.4',
+      '11112222:3333:4444:5555:6666:7777:8888',
+      '::1//64',
+      '::1/0001',
+      '1111:',
+      '1111:1.2.3.4',
+      '1111:2222',
+      '1111:22223333:4444:5555:6666:1.2.3.4',
+      '1111:22223333:4444:5555:6666:7777:8888',
+      '1111:2222:',
+      '1111:2222:1.2.3.4',
+      '1111:2222:3333',
+      '1111:2222:33334444:5555:6666:1.2.3.4',
+      '1111:2222:33334444:5555:6666:7777:8888',
+      '1111:2222:3333:',
+      '1111:2222:3333:1.2.3.4',
+      '1111:2222:3333:4444',
+      '1111:2222:3333:44445555:6666:1.2.3.4',
+      '1111:2222:3333:44445555:6666:7777:8888',
+      '1111:2222:3333:4444:',
+      '1111:2222:3333:4444:1.2.3.4',
+      '1111:2222:3333:4444:5555',
+      '1111:2222:3333:4444:55556666:1.2.3.4',
+      '1111:2222:3333:4444:55556666:7777:8888',
+      '1111:2222:3333:4444:5555:',
+      '1111:2222:3333:4444:5555:1.2.3.4',
+      '1111:2222:3333:4444:5555:6666',
+      '1111:2222:3333:4444:5555:66661.2.3.4',
+      '1111:2222:3333:4444:5555:66667777:8888',
+      '1111:2222:3333:4444:5555:6666:',
+      '1111:2222:3333:4444:5555:6666:1.2.3.4.5',
+      '1111:2222:3333:4444:5555:6666:255.255.255255',
+      '1111:2222:3333:4444:5555:6666:255.255255.255',
+      '1111:2222:3333:4444:5555:6666:255255.255.255',
+      '1111:2222:3333:4444:5555:6666:256.256.256.256',
+      '1111:2222:3333:4444:5555:6666:7777',
+      '1111:2222:3333:4444:5555:6666:77778888',
+      '1111:2222:3333:4444:5555:6666:7777:',
+      '1111:2222:3333:4444:5555:6666:7777:1.2.3.4',
+      '1111:2222:3333:4444:5555:6666:7777:::',
+      '1111:2222:3333:4444:5555:6666::8888:',
+      '1111:2222:3333:4444:5555:6666:::',
+      '1111:2222:3333:4444:5555:6666:::8888',
+      '1111:2222:3333:4444:5555::7777:8888:',
+      '1111:2222:3333:4444:5555::7777::',
+      '1111:2222:3333:4444:5555::8888:',
+      '1111:2222:3333:4444:5555:::',
+      '1111:2222:3333:4444:5555:::1.2.3.4',
+      '1111:2222:3333:4444:5555:::7777:8888',
+      '1111:2222:3333:4444::5555:',
+      '1111:2222:3333:4444::6666:7777:8888:',
+      '1111:2222:3333:4444::6666:7777::',
+      '1111:2222:3333:4444::6666::8888',
+      '1111:2222:3333:4444::7777:8888:',
+      '1111:2222:3333:4444::8888:',
+      '1111:2222:3333:4444:::',
+      '1111:2222:3333:4444:::6666:1.2.3.4',
+      '1111:2222:3333:4444:::6666:7777:8888',
+      '1111:2222:3333::5555:',
+      '1111:2222:3333::5555:6666:7777:8888:',
+      '1111:2222:3333::5555:6666:7777::',
+      '1111:2222:3333::5555:6666::8888',
+      '1111:2222:3333::5555::1.2.3.4',
+      '1111:2222:3333::5555::7777:8888',
+      '1111:2222:3333::6666:7777:8888:',
+      '1111:2222:3333::7777:8888:',
+      '1111:2222:3333::8888:',
+      '1111:2222:3333:::',
+      '1111:2222:3333:::5555:6666:1.2.3.4',
+      '1111:2222:3333:::5555:6666:7777:8888',
+      '1111:2222::4444:5555:6666:7777:8888:',
+      '1111:2222::4444:5555:6666:7777::',
+      '1111:2222::4444:5555:6666::8888',
+      '1111:2222::4444:5555::1.2.3.4',
+      '1111:2222::4444:5555::7777:8888',
+      '1111:2222::4444::6666:1.2.3.4',
+      '1111:2222::4444::6666:7777:8888',
+      '1111:2222::5555:',
+      '1111:2222::5555:6666:7777:8888:',
+      '1111:2222::6666:7777:8888:',
+      '1111:2222::7777:8888:',
+      '1111:2222::8888:',
+      '1111:2222:::',
+      '1111:2222:::4444:5555:6666:1.2.3.4',
+      '1111:2222:::4444:5555:6666:7777:8888',
+      '1111::3333:4444:5555:6666:7777:8888:',
+      '1111::3333:4444:5555:6666:7777::',
+      '1111::3333:4444:5555:6666::8888',
+      '1111::3333:4444:5555::1.2.3.4',
+      '1111::3333:4444:5555::7777:8888',
+      '1111::3333:4444::6666:1.2.3.4',
+      '1111::3333:4444::6666:7777:8888',
+      '1111::3333::5555:6666:1.2.3.4',
+      '1111::3333::5555:6666:7777:8888',
+      '1111::4444:5555:6666:7777:8888:',
+      '1111::5555:',
+      '1111::5555:6666:7777:8888:',
+      '1111::6666:7777:8888:',
+      '1111::7777:8888:',
+      '1111::8888:',
+      '1111:::',
+      '1111:::3333:4444:5555:6666:1.2.3.4',
+      '1111:::3333:4444:5555:6666:7777:8888',
+      '12345::6:7:8',
+      '124.15.6.89/60',
+      '1:2:3:4:5:6:7:8:9',
+      '1:2:3::4:5:6:7:8:9',
+      '1:2:3::4:5::7:8',
+      '1::1.2.256.4',
+      '1::1.2.3.256',
+      '1::1.2.3.300',
+      '1::1.2.3.900',
+      '1::1.2.300.4',
+      '1::1.2.900.4',
+      '1::1.256.3.4',
+      '1::1.300.3.4',
+      '1::1.900.3.4',
+      '1::256.2.3.4',
+      '1::260.2.3.4',
+      '1::2::3',
+      '1::300.2.3.4',
+      '1::300.300.300.300',
+      '1::3000.30.30.30',
+      '1::400.2.3.4',
+      '1::5:1.2.256.4',
+      '1::5:1.2.3.256',
+      '1::5:1.2.3.300',
+      '1::5:1.2.3.900',
+      '1::5:1.2.300.4',
+      '1::5:1.2.900.4',
+      '1::5:1.256.3.4',
+      '1::5:1.300.3.4',
+      '1::5:1.900.3.4',
+      '1::5:256.2.3.4',
+      '1::5:260.2.3.4',
+      '1::5:300.2.3.4',
+      '1::5:300.300.300.300',
+      '1::5:3000.30.30.30',
+      '1::5:400.2.3.4',
+      '1::5:900.2.3.4',
+      '1::900.2.3.4',
+      '1:::3:4:5',
+      '2001:0000:1234: 0000:0000:C1C0:ABCD:0876',
+      '2001:0000:1234:0000:0000:C1C0:ABCD:0876  0',
+      '2001:1:1:1:1:1:255Z255X255Y255',
+      '2001::FFD3::57ab',
+      '2001:DB8:0:0:8:800:200C:417A:221',
+      '2001:db8:85a3::8a2e:37023:7334',
+      '2001:db8:85a3::8a2e:370k:7334',
+      '3ffe:0b00:0000:0001:0000:0000:000a',
+      '3ffe:b00::1::a',
+      ':',
+      ':1.2.3.4',
+      ':1111:2222:3333:4444:5555:6666:1.2.3.4',
+      ':1111:2222:3333:4444:5555:6666:7777:8888',
+      ':1111:2222:3333:4444:5555:6666:7777::',
+      ':1111:2222:3333:4444:5555:6666::',
+      ':1111:2222:3333:4444:5555:6666::8888',
+      ':1111:2222:3333:4444:5555::',
+      ':1111:2222:3333:4444:5555::1.2.3.4',
+      ':1111:2222:3333:4444:5555::7777:8888',
+      ':1111:2222:3333:4444:5555::8888',
+      ':1111:2222:3333:4444::',
+      ':1111:2222:3333:4444::1.2.3.4',
+      ':1111:2222:3333:4444::5555',
+      ':1111:2222:3333:4444::6666:1.2.3.4',
+      ':1111:2222:3333:4444::6666:7777:8888',
+      ':1111:2222:3333:4444::7777:8888',
+      ':1111:2222:3333:4444::8888',
+      ':1111:2222:3333::',
+      ':1111:2222:3333::1.2.3.4',
+      ':1111:2222:3333::5555',
+      ':1111:2222:3333::5555:6666:1.2.3.4',
+      ':1111:2222:3333::5555:6666:7777:8888',
+      ':1111:2222:3333::6666:1.2.3.4',
+      ':1111:2222:3333::6666:7777:8888',
+      ':1111:2222:3333::7777:8888',
+      ':1111:2222:3333::8888',
+      ':1111:2222::',
+      ':1111:2222::1.2.3.4',
+      ':1111:2222::4444:5555:6666:1.2.3.4',
+      ':1111:2222::4444:5555:6666:7777:8888',
+      ':1111:2222::5555',
+      ':1111:2222::5555:6666:1.2.3.4',
+      ':1111:2222::5555:6666:7777:8888',
+      ':1111:2222::6666:1.2.3.4',
+      ':1111:2222::6666:7777:8888',
+      ':1111:2222::7777:8888',
+      ':1111:2222::8888',
+      ':1111::',
+      ':1111::1.2.3.4',
+      ':1111::3333:4444:5555:6666:1.2.3.4',
+      ':1111::3333:4444:5555:6666:7777:8888',
+      ':1111::4444:5555:6666:1.2.3.4',
+      ':1111::4444:5555:6666:7777:8888',
+      ':1111::5555',
+      ':1111::5555:6666:1.2.3.4',
+      ':1111::5555:6666:7777:8888',
+      ':1111::6666:1.2.3.4',
+      ':1111::6666:7777:8888',
+      ':1111::7777:8888',
+      ':1111::8888',
+      ':2222:3333:4444:5555:6666:1.2.3.4',
+      ':2222:3333:4444:5555:6666:7777:8888',
+      ':3333:4444:5555:6666:1.2.3.4',
+      ':3333:4444:5555:6666:7777:8888',
+      ':4444:5555:6666:1.2.3.4',
+      ':4444:5555:6666:7777:8888',
+      ':5555:6666:1.2.3.4',
+      ':5555:6666:7777:8888',
+      ':6666:1.2.3.4',
+      ':6666:7777:8888',
+      ':7777:8888',
+      ':8888',
+      '::-1',
+      '::.',
+      '::..',
+      '::...',
+      '::...4',
+      '::..3.',
+      '::..3.4',
+      '::.2..',
+      '::.2.3.',
+      '::.2.3.4',
+      '::1...',
+      '::1.2..',
+      '::1.2.256.4',
+      '::1.2.3.',
+      '::1.2.3.256',
+      '::1.2.3.300',
+      '::1.2.3.900',
+      '::1.2.300.4',
+      '::1.2.900.4',
+      '::1.256.3.4',
+      '::1.300.3.4',
+      '::1.900.3.4',
+      '::1111:2222:3333:4444:5555:6666::',
+      '::2222:3333:4444:5555:6666:7777:8888:',
+      '::2222:3333:4444:5555:7777:8888::',
+      '::2222:3333:4444:5555:7777::8888',
+      '::2222:3333:4444:5555::1.2.3.4',
+      '::2222:3333:4444:5555::7777:8888',
+      '::2222:3333:4444::6666:1.2.3.4',
+      '::2222:3333:4444::6666:7777:8888',
+      '::2222:3333::5555:6666:1.2.3.4',
+      '::2222:3333::5555:6666:7777:8888',
+      '::2222::4444:5555:6666:1.2.3.4',
+      '::2222::4444:5555:6666:7777:8888',
+      '::256.2.3.4',
+      '::260.2.3.4',
+      '::300.2.3.4',
+      '::300.300.300.300',
+      '::3000.30.30.30',
+      '::3333:4444:5555:6666:7777:8888:',
+      '::400.2.3.4',
+      '::4444:5555:6666:7777:8888:',
+      '::5555:',
+      '::5555:6666:7777:8888:',
+      '::6666:7777:8888:',
+      '::7777:8888:',
+      '::8888:',
+      '::900.2.3.4',
+      ':::',
+      ':::1.2.3.4',
+      ':::2222:3333:4444:5555:6666:1.2.3.4',
+      ':::2222:3333:4444:5555:6666:7777:8888',
+      ':::3333:4444:5555:6666:7777:8888',
+      ':::4444:5555:6666:1.2.3.4',
+      ':::4444:5555:6666:7777:8888',
+      ':::5555',
+      ':::5555:6666:1.2.3.4',
+      ':::5555:6666:7777:8888',
+      ':::6666:1.2.3.4',
+      ':::6666:7777:8888',
+      ':::7777:8888',
+      ':::8888',
+      '::ffff:192x168.1.26',
+      '::ffff:2.3.4',
+      '::ffff:257.1.2.3',
+      'FF01::101::2',
+      'FF02:0000:0000:0000:0000:0000:0000:0000:0001',
+      'XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:1.2.3.4',
+      'XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:XXXX',
+      'a::b::c',
+      'a::g',
+      'a:a:a:a:a:a:a:a:a',
+      'a:aaaaa::',
+      'a:b',
+      'a:b:c:d:e:f:g:0',
+      'ffff:',
+      'ffff::ffff::ffff',
+      'ffgg:ffff:ffff:ffff:ffff:ffff:ffff:ffff',
+      'ldkfj',
+      '::/129',
+      '1000:://32',
+      '::/',
     ]
-    invalidIpV6.map(ip=>{
-      const valid=  isValidIP(ip)
+    invalidIpV6.map((ip) => {
+      const valid = isValidIP(ip)
       assert.equal(valid, false)
     })
   })
 
   it('Check list of IPV6 Valid addresses', () => {
-    const validIpv6 =[
-      "0000:0000:0000:0000:0000:0000:0000:0000",
-      "0000:0000:0000:0000:0000:0000:0000:0001",
-      "0:0:0:0:0:0:0:0",
-      "0:0:0:0:0:0:0:1",
-      "0:0:0:0:0:0:0::",
-      "0:0:0:0:0:0:13.1.68.3",
-      "0:0:0:0:0:0::",
-      "0:0:0:0:0::",
-      "0:0:0:0:0:FFFF:129.144.52.38",
-      "0:0:0:0:1:0:0:0",
-      "0:0:0:0::",
-      "0:0:0::",
-      "0:0::",
-      "0:1:2:3:4:5:6:7",
-      "0::",
-      "0:a:b:c:d:e:f::",
-      "1080:0:0:0:8:800:200c:417a",
-      "1080::8:800:200c:417a",
-      "1111:2222:3333:4444:5555:6666:123.123.123.123",
-      "1111:2222:3333:4444:5555:6666:7777:8888",
-      "1111:2222:3333:4444:5555:6666:7777::",
-      "1111:2222:3333:4444:5555:6666::",
-      "1111:2222:3333:4444:5555:6666::8888",
-      "1111:2222:3333:4444:5555::",
-      "1111:2222:3333:4444:5555::7777:8888",
-      "1111:2222:3333:4444:5555::8888",
-      "1111:2222:3333:4444::",
-      "1111:2222:3333:4444::6666:123.123.123.123",
-      "1111:2222:3333:4444::6666:7777:8888",
-      "1111:2222:3333:4444::7777:8888",
-      "1111:2222:3333:4444::8888",
-      "1111:2222:3333::",
-      "1111:2222:3333::5555:6666:123.123.123.123",
-      "1111:2222:3333::5555:6666:7777:8888",
-      "1111:2222:3333::6666:123.123.123.123",
-      "1111:2222:3333::6666:7777:8888",
-      "1111:2222:3333::7777:8888",
-      "1111:2222:3333::8888",
-      "1111:2222::",
-      "1111:2222::4444:5555:6666:123.123.123.123",
-      "1111:2222::4444:5555:6666:7777:8888",
-      "1111:2222::5555:6666:123.123.123.123",
-      "1111:2222::5555:6666:7777:8888",
-      "1111:2222::6666:123.123.123.123",
-      "1111:2222::6666:7777:8888",
-      "1111:2222::7777:8888",
-      "1111:2222::8888",
-      "1111::",
-      "1111::3333:4444:5555:6666:123.123.123.123",
-      "1111::3333:4444:5555:6666:7777:8888",
-      "1111::4444:5555:6666:123.123.123.123",
-      "1111::4444:5555:6666:7777:8888",
-      "1111::5555:6666:123.123.123.123",
-      "1111::5555:6666:7777:8888",
-      "1111::6666:123.123.123.123",
-      "1111::6666:7777:8888",
-      "1111::7777:8888",
-      "1111::8888",
-      "1:2:3:4:5:6:1.2.3.4",
-      "1:2:3:4:5:6:7:8",
-      "1:2:3:4:5:6::",
-      "1:2:3:4:5:6::8",
-      "1:2:3:4:5::",
-      "1:2:3:4:5::7:8",
-      "1:2:3:4:5::8",
-      "1:2:3:4::",
-      "1:2:3:4::5:1.2.3.4",
-      "1:2:3:4::7:8",
-      "1:2:3:4::8",
-      "1:2:3::",
-      "1:2:3::5:1.2.3.4",
-      "1:2:3::7:8",
-      "1:2:3::8",
-      "1:2::",
-      "1:2::5:1.2.3.4",
-      "1:2::7:8",
-      "1:2::8",
-      "1::",
-      "1::2:3",
-      "1::2:3:4",
-      "1::2:3:4:5",
-      "1::2:3:4:5:6",
-      "1::2:3:4:5:6:7",
-      "1::5:1.2.3.4",
-      "1::5:11.22.33.44",
-      "1::7:8",
-      "1::8",
-      "2001:0000:1234:0000:0000:C1C0:ABCD:0876",
-      "2001:0000:4136:e378:8000:63bf:3fff:fdd2",
-      "2001:0db8:0000:0000:0000:0000:1428:57ab",
-      "2001:0db8:0000:0000:0000::1428:57ab",
-      "2001:0db8:0:0:0:0:1428:57ab",
-      "2001:0db8:0:0::1428:57ab",
-      "2001:0db8:1234:0000:0000:0000:0000:0000",
-      "2001:0db8:1234::",
-      "2001:0db8:1234:ffff:ffff:ffff:ffff:ffff",
-      "2001:0db8:85a3:0000:0000:8a2e:0370:7334",
-      "2001:0db8::1428:57ab",
-      "2001::CE49:7601:2CAD:DFFF:7C94:FFFE",
-      "2001::CE49:7601:E866:EFFF:62C3:FFFE",
-      "2001:DB8:0:0:8:800:200C:417A",
-      "2001:DB8::8:800:200C:417A",
-      "2001:db8:85a3:0:0:8a2e:370:7334",
-      "2001:db8:85a3::8a2e:370:7334",
-      "2001:db8::",
-      "2001:db8::1428:57ab",
-      "2001:db8:a::123",
-      "2002::",
-      "2608::3:5",
-      "2608:af09:30:0:0:0:0:134",
-      "2608:af09:30::102a:7b91:c239:baff",
-      "2::10",
-      "3ffe:0b00:0000:0000:0001:0000:0000:000a",
-      "7:6:5:4:3:2:1:0",
-      "::",
-      "::0",
-      "::0:0",
-      "::0:0:0",
-      "::0:0:0:0",
-      "::0:0:0:0:0",
-      "::0:0:0:0:0:0",
-      "::0:0:0:0:0:0:0",
-      "::0:a:b:c:d:e:f",
-      "::1",
-      "::123.123.123.123",
-      "::13.1.68.3",
-      "::2222:3333:4444:5555:6666:123.123.123.123",
-      "::2222:3333:4444:5555:6666:7777:8888",
-      "::2:3",
-      "::2:3:4",
-      "::2:3:4:5",
-      "::2:3:4:5:6",
-      "::2:3:4:5:6:7",
-      "::2:3:4:5:6:7:8",
-      "::3333:4444:5555:6666:7777:8888",
-      "::4444:5555:6666:123.123.123.123",
-      "::4444:5555:6666:7777:8888",
-      "::5555:6666:123.123.123.123",
-      "::5555:6666:7777:8888",
-      "::6666:123.123.123.123",
-      "::6666:7777:8888",
-      "::7777:8888",
-      "::8",
-      "::8888",
-      "::FFFF:129.144.52.38",
-      "::ffff:0:0",
-      "::ffff:0c22:384e",
-      "::ffff:12.34.56.78",
-      "::ffff:192.0.2.128",
-      "::ffff:192.168.1.1",
-      "::ffff:192.168.1.26",
-      "::ffff:c000:280",
-      "FF01:0:0:0:0:0:0:101",
-      "FF01::101",
-      "FF02:0000:0000:0000:0000:0000:0000:0001",
-      "a:b:c:d:e:f:0::",
-      "fe80:0000:0000:0000:0204:61ff:fe9d:f156",
-      "fe80:0:0:0:204:61ff:254.157.241.86",
-      "fe80:0:0:0:204:61ff:fe9d:f156",
-      "fe80::",
-      "fe80::1",
-      "fe80::204:61ff:254.157.241.86",
-      "fe80::204:61ff:fe9d:f156",
-      "fe80::217:f2ff:254.7.237.98",
-      "fe80::217:f2ff:fe07:ed62",
-      "fedc:ba98:7654:3210:fedc:ba98:7654:3210",
-      "ff02::1",
-      "ffff::",
-      "ffff::3:5",
-      "ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff",
-      "a:0::0:b",
-      "a:0:0::0:b",
-      "a:0::0:0:b",
-      "a::0:0:b",
-      "a::0:b",
-      "a:0::b",
-      "a:0:0::b"
+    const validIpv6 = [
+      '0000:0000:0000:0000:0000:0000:0000:0000',
+      '0000:0000:0000:0000:0000:0000:0000:0001',
+      '0:0:0:0:0:0:0:0',
+      '0:0:0:0:0:0:0:1',
+      '0:0:0:0:0:0:0::',
+      '0:0:0:0:0:0:13.1.68.3',
+      '0:0:0:0:0:0::',
+      '0:0:0:0:0::',
+      '0:0:0:0:0:FFFF:129.144.52.38',
+      '0:0:0:0:1:0:0:0',
+      '0:0:0:0::',
+      '0:0:0::',
+      '0:0::',
+      '0:1:2:3:4:5:6:7',
+      '0::',
+      '0:a:b:c:d:e:f::',
+      '1080:0:0:0:8:800:200c:417a',
+      '1080::8:800:200c:417a',
+      '1111:2222:3333:4444:5555:6666:123.123.123.123',
+      '1111:2222:3333:4444:5555:6666:7777:8888',
+      '1111:2222:3333:4444:5555:6666:7777::',
+      '1111:2222:3333:4444:5555:6666::',
+      '1111:2222:3333:4444:5555:6666::8888',
+      '1111:2222:3333:4444:5555::',
+      '1111:2222:3333:4444:5555::7777:8888',
+      '1111:2222:3333:4444:5555::8888',
+      '1111:2222:3333:4444::',
+      '1111:2222:3333:4444::6666:123.123.123.123',
+      '1111:2222:3333:4444::6666:7777:8888',
+      '1111:2222:3333:4444::7777:8888',
+      '1111:2222:3333:4444::8888',
+      '1111:2222:3333::',
+      '1111:2222:3333::5555:6666:123.123.123.123',
+      '1111:2222:3333::5555:6666:7777:8888',
+      '1111:2222:3333::6666:123.123.123.123',
+      '1111:2222:3333::6666:7777:8888',
+      '1111:2222:3333::7777:8888',
+      '1111:2222:3333::8888',
+      '1111:2222::',
+      '1111:2222::4444:5555:6666:123.123.123.123',
+      '1111:2222::4444:5555:6666:7777:8888',
+      '1111:2222::5555:6666:123.123.123.123',
+      '1111:2222::5555:6666:7777:8888',
+      '1111:2222::6666:123.123.123.123',
+      '1111:2222::6666:7777:8888',
+      '1111:2222::7777:8888',
+      '1111:2222::8888',
+      '1111::',
+      '1111::3333:4444:5555:6666:123.123.123.123',
+      '1111::3333:4444:5555:6666:7777:8888',
+      '1111::4444:5555:6666:123.123.123.123',
+      '1111::4444:5555:6666:7777:8888',
+      '1111::5555:6666:123.123.123.123',
+      '1111::5555:6666:7777:8888',
+      '1111::6666:123.123.123.123',
+      '1111::6666:7777:8888',
+      '1111::7777:8888',
+      '1111::8888',
+      '1:2:3:4:5:6:1.2.3.4',
+      '1:2:3:4:5:6:7:8',
+      '1:2:3:4:5:6::',
+      '1:2:3:4:5:6::8',
+      '1:2:3:4:5::',
+      '1:2:3:4:5::7:8',
+      '1:2:3:4:5::8',
+      '1:2:3:4::',
+      '1:2:3:4::5:1.2.3.4',
+      '1:2:3:4::7:8',
+      '1:2:3:4::8',
+      '1:2:3::',
+      '1:2:3::5:1.2.3.4',
+      '1:2:3::7:8',
+      '1:2:3::8',
+      '1:2::',
+      '1:2::5:1.2.3.4',
+      '1:2::7:8',
+      '1:2::8',
+      '1::',
+      '1::2:3',
+      '1::2:3:4',
+      '1::2:3:4:5',
+      '1::2:3:4:5:6',
+      '1::2:3:4:5:6:7',
+      '1::5:1.2.3.4',
+      '1::5:11.22.33.44',
+      '1::7:8',
+      '1::8',
+      '2001:0000:1234:0000:0000:C1C0:ABCD:0876',
+      '2001:0000:4136:e378:8000:63bf:3fff:fdd2',
+      '2001:0db8:0000:0000:0000:0000:1428:57ab',
+      '2001:0db8:0000:0000:0000::1428:57ab',
+      '2001:0db8:0:0:0:0:1428:57ab',
+      '2001:0db8:0:0::1428:57ab',
+      '2001:0db8:1234:0000:0000:0000:0000:0000',
+      '2001:0db8:1234::',
+      '2001:0db8:1234:ffff:ffff:ffff:ffff:ffff',
+      '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
+      '2001:0db8::1428:57ab',
+      '2001::CE49:7601:2CAD:DFFF:7C94:FFFE',
+      '2001::CE49:7601:E866:EFFF:62C3:FFFE',
+      '2001:DB8:0:0:8:800:200C:417A',
+      '2001:DB8::8:800:200C:417A',
+      '2001:db8:85a3:0:0:8a2e:370:7334',
+      '2001:db8:85a3::8a2e:370:7334',
+      '2001:db8::',
+      '2001:db8::1428:57ab',
+      '2001:db8:a::123',
+      '2002::',
+      '2608::3:5',
+      '2608:af09:30:0:0:0:0:134',
+      '2608:af09:30::102a:7b91:c239:baff',
+      '2::10',
+      '3ffe:0b00:0000:0000:0001:0000:0000:000a',
+      '7:6:5:4:3:2:1:0',
+      '::',
+      '::0',
+      '::0:0',
+      '::0:0:0',
+      '::0:0:0:0',
+      '::0:0:0:0:0',
+      '::0:0:0:0:0:0',
+      '::0:0:0:0:0:0:0',
+      '::0:a:b:c:d:e:f',
+      '::1',
+      '::123.123.123.123',
+      '::13.1.68.3',
+      '::2222:3333:4444:5555:6666:123.123.123.123',
+      '::2222:3333:4444:5555:6666:7777:8888',
+      '::2:3',
+      '::2:3:4',
+      '::2:3:4:5',
+      '::2:3:4:5:6',
+      '::2:3:4:5:6:7',
+      '::2:3:4:5:6:7:8',
+      '::3333:4444:5555:6666:7777:8888',
+      '::4444:5555:6666:123.123.123.123',
+      '::4444:5555:6666:7777:8888',
+      '::5555:6666:123.123.123.123',
+      '::5555:6666:7777:8888',
+      '::6666:123.123.123.123',
+      '::6666:7777:8888',
+      '::7777:8888',
+      '::8',
+      '::8888',
+      '::FFFF:129.144.52.38',
+      '::ffff:0:0',
+      '::ffff:0c22:384e',
+      '::ffff:12.34.56.78',
+      '::ffff:192.0.2.128',
+      '::ffff:192.168.1.1',
+      '::ffff:192.168.1.26',
+      '::ffff:c000:280',
+      'FF01:0:0:0:0:0:0:101',
+      'FF01::101',
+      'FF02:0000:0000:0000:0000:0000:0000:0001',
+      'a:b:c:d:e:f:0::',
+      'fe80:0000:0000:0000:0204:61ff:fe9d:f156',
+      'fe80:0:0:0:204:61ff:254.157.241.86',
+      'fe80:0:0:0:204:61ff:fe9d:f156',
+      'fe80::',
+      'fe80::1',
+      'fe80::204:61ff:254.157.241.86',
+      'fe80::204:61ff:fe9d:f156',
+      'fe80::217:f2ff:254.7.237.98',
+      'fe80::217:f2ff:fe07:ed62',
+      'fedc:ba98:7654:3210:fedc:ba98:7654:3210',
+      'ff02::1',
+      'ffff::',
+      'ffff::3:5',
+      'ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff',
+      'a:0::0:b',
+      'a:0:0::0:b',
+      'a:0::0:0:b',
+      'a::0:0:b',
+      'a::0:b',
+      'a:0::b',
+      'a:0:0::b',
     ]
-    validIpv6.map(ip=>{
-      const valid=  isValidIP(ip)
+    validIpv6.map((ip) => {
+      const valid = isValidIP(ip)
       assert.equal(valid, true)
-
     })
   })
 })
